### 2022-09-27 #4803 @editorialbot
[PRE REVIEW]: X-PSI: A Python package for neutron star X-ray Pulse Simulation and Inference
**Submitting author:** @drannawatts (<a href="http://orcid.org/0000-0002-1009-2354">Anna Watts</a>)
**Repository:** https://github.com/xpsi-group/xpsi
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
