### 2022-09-29 #4803 @drannawatts
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/3136749595?check_suite_focus=true).

We think we have identified the issue (with an ORCID) and fixed it.  That said, the JOSS Github action managed to generate a pdf for us on the 1st version without a problem!
