### 2022-10-10 2db1fd5
fixing input bugs and typo
