### 2022-10-10 39aaa6c
fixing spaces in ProjectionTool.py
