### 2022-10-18 428c4ec
Initial script for automatic deployment of docs
