### 2022-10-18 ba97c72
Commented out linting in CI tests for now until Py3 is up and running
