+++ .github/workflows/ci_tests.yml
@@ -32,13 +32,14 @@ jobs:
     - name: install gsl
       run: |
         sudo apt-get install -y libgsl-dev
-    - name: Lint with flake8
-      run: |
-        conda install flake8
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        #flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+# commented out linting until Py3 version is in main branch
+#    - name: Lint with flake8
+#      run: |
+#        conda install flake8
+#        # stop the build if there are Python syntax errors or undefined names
+#        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+#        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
+#        #flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: install X-PSI package
       run: |
         python setup.py install
