### 2022-10-18 ba69818
Don't build JOSS every time
