+++ .github/workflows/build_docs.yml
@@ -1,10 +1,9 @@
 name: Build docs
 
-
 on:
   push:
-#    branches:
-#      - "main"
+    branches:
+      - "main"
   pull_request:
   # Allows you to run this workflow manually from the Actions tab
   workflow_dispatch:
@@ -50,6 +49,7 @@ jobs:
     - name : build docs
       run: |
         conda install sphinx 
+        conda install -c conda-forge nbsphinx
         make html
       working-directory: docs/
     - name: Upload artifact
