### 2022-10-19 9be985a
Build sphinx with specific package versions to make it work on Python2
