+++ .github/workflows/build_docs.yml
@@ -48,9 +48,11 @@ jobs:
         CC:   gcc-10
     - name : install docs dependencies
       run: |
-        conda install sphinx 
-        conda install -c conda-forge nbsphinx
-        pip install sphinx_rtd_theme
+        conda install sphinx=1.8.5
+        conda install -c conda-forge nbsphinx=0.5.1
+        conda install decorator=4.4.1
+        pip install sphinxcontrib-websupport==1.1.2
+        pip install sphinx_rtd_theme==0.4.3
     - name : build docs
       run: make html
       working-directory: docs/
