### 2022-10-19 4add628
Another path solving attmept
