+++ .github/workflows/build_docs.yml
@@ -33,6 +33,13 @@ jobs:
          environment-file: basic_environment.yml
          python-version: 2.7
          auto-activate-base: false
+    - name : install docs dependencies
+      run: |
+        conda install sphinx=1.8.5
+        conda install -c conda-forge nbsphinx=0.5.1
+        conda install decorator=4.4.1
+        pip install sphinxcontrib-websupport==1.1.2
+        pip install sphinx_rtd_theme==0.4.3
     - name: Set up GCC
       run: |
          sudo apt-get install -y gcc
@@ -46,18 +53,12 @@ jobs:
       shell: bash -l {0}
       env:
         CC:   gcc-10
-    - name : install docs dependencies
-      run: |
-        conda install sphinx=1.8.5
-        conda install -c conda-forge nbsphinx=0.5.1
-        conda install decorator=4.4.1
-        pip install sphinxcontrib-websupport==1.1.2
-        pip install sphinx_rtd_theme==0.4.3
     - name : build docs
       run: make html
       working-directory: docs/
     - name: Upload artifact
       uses: actions/upload-pages-artifact@v1
+      path: docs/ 
     - name: Deploy to GitHub Pages
       if: success()
       uses: crazy-max/ghaction-github-pages@v2
