### 2022-10-19 29b4795
removed artifact step
