### 2022-10-19 e7dd42a
Try without explicit settings of permissions
