+++ .github/workflows/build_docs.yml
@@ -10,17 +10,17 @@ on:
 
 
 # Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
-permissions:
-  contents: read
-  pages: write
-  id-token: write
+#permissions:
+#  contents: read
+#  pages: write
+#  id-token: write
 
 jobs:
   build:
     runs-on: ubuntu-latest
-#    defaults:
-#      run:
-#        shell: bash -l {0}
+    defaults:
+      run:
+        shell: bash -l {0}
     strategy:
       fail-fast: false
       max-parallel: 5
