### 2022-10-19 61ee76b
Fixed error in using the shell
