+++ .github/workflows/build_docs.yml
@@ -40,7 +40,6 @@ jobs:
         conda install decorator=4.4.1
         pip install sphinxcontrib-websupport==1.1.2
         pip install sphinx_rtd_theme==0.4.3
-        shell: bash -l {0}
     - name: Set up GCC
       run: |
          sudo apt-get install -y gcc
