### 2022-10-19 a6bd40e
Removed change, cleaned up workflow script
