+++ .github/workflows/build_docs.yml
@@ -3,17 +3,9 @@ name: Build docs
 on:
   push:
     branches:
-      - "main"
-  pull_request:
-  # Allows you to run this workflow manually from the Actions tab
+      - main
+  release:
   workflow_dispatch:
-
-
-# Sets permissions of the GITHUB_TOKEN to allow deployment to GitHub Pages
-#permissions:
-#  contents: read
-#  pages: write
-#  id-token: write
 
 jobs:
   build:
