### 2022-11-02 814e0e0
Fixes #175 issue with parsing values using `module_generator`
