+++ xpsi/module_generator.py
@@ -1394,26 +1394,6 @@ def parse_bounds(bounds, value, default_to_free=True):
 
     return None
 
-bounds = dict(neutral_hydrogen_column_density = parse_bounds(args.neutral_hydrogen_column_density_bounds,
-                                                              args.neutral_hydrogen_column_density_value))
-values = dict(neutral_hydrogen_column_density = args.neutral_hydrogen_column_density_value)
-
-interstellar = CustomInterstellar.load(args.attenuation_path,
-                                       args.attenuation_energy_column,
-                                       args.attenuation_column,
-                                       bounds = bounds,
-                                       values = values)
-'''
-)
-
-module += (
-'''
-signals = [[],]
-'''
-)
-
-module += (
-'''
 def derived_parameter(func, parameter, space='caller'):
 
     class derive(Derive):
@@ -1434,6 +1414,26 @@ def parse_value(value):
             return derived_parameter(*eval(value))
     else:
         return None
+'''
+)
+
+module += (
+'''
+bounds = dict(neutral_hydrogen_column_density = parse_bounds(args.neutral_hydrogen_column_density_bounds,
+                                                              args.neutral_hydrogen_column_density_value))
+values = dict(neutral_hydrogen_column_density = args.neutral_hydrogen_column_density_value)
+
+interstellar = CustomInterstellar.load(args.attenuation_path,
+                                       args.attenuation_energy_column,
+                                       args.attenuation_column,
+                                       bounds = bounds,
+                                       values = values)
+'''
+)
+
+module += (
+'''
+signals = [[],]
 '''
 )
 
