### 2022-11-09 8c73f73
Added utilities directories with ImportanceSample.py and ProjectionTool.py (#145)
* adding utilities directory with ProjectionTool and ImportanceSample script

* added return ax in ProjectionTool

* Update ProjectionTool.py

* Update ImportanceSample.py

* fixing input bugs and typo

* Other bugs and typos

* correcting space in ProjectTool

* Other bugs in inputs Importance sampling

* fixing spaces in ProjectionTool.py

* Update setup.py

* Create __init__.py

* changing alignment on lines

Co-authored-by: Serena Vinciguerra <s.vinciguerra@uva.nl>
