+++ xpsi/utilities/ImportanceSample.py
@@ -1,0 +1,131 @@
+from __future__ import print_function, division
+import sys
+import argparse
+
+
+print ("WARNING: Please be aware that ERRORS can arise if relative paths are use in the main.py file (to point to the config file) or in the config.ini file (to point to other files used in the analysis).")
+
+parser = argparse.ArgumentParser(
+    description='''
+    Main module for X-PSI Importance Sample.
+    It requires a file main_IS.py in the same directory as the main.py where the importance sampling settings are set. For changes in the priors, this need to include importing a different CustomPrior, e.g. CustomPrior_IS.py in the main and the presence in both CustomPrior.py and the CustomPrior_IS.py of the method density, to evaluate prior density functions. 
+    ''',
+    fromfile_prefix_chars='@')
+
+parser.add_argument('--name-original-main', type=str, help='Name of original main file')
+parser.add_argument('--name-importance-sample-main', type=str, help='Name of importance sample main file')
+parser.add_argument('--path-to-mains', type=str, help='Absolute or relative path to main files')
+parser.add_argument('--path-to-outputs', type=str, help='Absolute or relative path to output files')
+parser.add_argument('--output-root-name', type=str, help='Name of output root file, no extension')
+parser.add_argument('--names', nargs='+',help='List of names of model variables')
+parser.add_argument('--likelihood-change', type=str, help='Importance sample for chaning likelihood')
+parser.add_argument('--prior-change', type=str, help='Importance sample for chaning priors')
+parser.add_argument('--weight-threshold', type=float, help='Weight threshold for importance sampling')
+parser.add_argument('--overwrite', type=str, help='Overwrite')
+
+args = parser.parse_args()
+
+if not args.path_to_mains:
+    args.path_to_mains = raw_input('Specify Absolute or relative path to main files: ')
+
+if not args.name_original_main:
+    args.name_original_main = raw_input('Specify name of original main file: ')
+    
+if not args.name_importance_sample_main:
+    args.name_importance_sample_main = raw_input('Specify name of importance sample main file: ')
+
+if not args.path_to_outputs:
+    args.path_to_outputs = raw_input('Specify Absolute or relative path to output files: ')
+    if args.path_to_outputs[-1] =='/':
+        args.path_to_outputs = args.path_to_outputs[:-1]
+        
+if not args.output_root_name:
+    args.output_root_name = raw_input('Specify name of output root file, without extension: ')
+
+if not args.names:
+    names = input('Specify the name of the model variables: ') #check if input
+    args.names = names.split()
+else:
+    args.names = args.names[0].split()
+
+def check_prior_change(test):
+    if not args.prior_change:
+        args.prior_change = raw_input('Specify if importance sampling changes priors of model variables  (True/False): ')
+        
+    if args.prior_change.capitalize() == 'False':
+        args.prior_change=False
+        test = False
+        return test
+    elif args.prior_change.capitalize() == 'True':
+        args.prior_change=True
+        test = False
+        return test
+    else:
+        print ('Only True or False options')
+        args.prior_change = raw_input('Specify if importance sampling changes priors of model variables  (True/False): ')
+        
+def check_likelihood_change(test):
+    if not args.likelihood_change:
+        args.likelihood_change = raw_input('Specify if importance sampling changes likelihood of model variables  (True/False): ')
+        
+    if args.likelihood_change.capitalize() == 'False':
+        args.likelihood_change=False
+        test = False
+        return test
+    elif args.likelihood_change.capitalize() == 'True':
+        args.likelihood_change=True
+        test = False
+        return test
+    else:
+        print ('Only True or False options')
[…]
