### 2022-11-09 4d9a472
Changelog updated and version changed to v1.1.
