+++ CHANGELOG.rst
@@ -34,7 +34,35 @@ and this project adheres to
 .. ^^^^^^^^^^^
 
 
-[v1.0] - 2022-09-26
+[v1.1.0] - 2022-11-09
+~~~~~~~~~~~~~~~~~~~
+
+Summary
+^^^^^^^
+
+* Additional tools included in a ``xpsi/utilities`` directory for plotting hot regions on a sphere and performing importance sampling in X-PSI. Documentation for these tools is to be appended later.
+
+Added
+^^^^^
+
+* ``xpsi/utilities/ProjectionTool.py``: Tool for projecting hot regions.
+
+* ``xpsi/utilities/ImportanceSample.py``: Tool for calling X-PSI importance sampling.
+
+Changed
+^^^^^^^
+
+* ``Setup.py`` file changed to include the new utilities directory.
+
+Attribution
+^^^^^^^^^^^
+
+Serena Vinciguerra,
+Tuomo Salmi,
+Devarshi Choudhury
+
+
+[v1.0.0] - 2022-09-26
 ~~~~~~~~~~~~~~~~~~~
 
 Summary
@@ -51,7 +79,7 @@ Added
 
 * A modification of the ``setup.py`` with flags (``--NumHot`` and ``--NumElse``) now facilitates switching between surface emission models.
 
-* The post-processing module now shows the credible intervals of each parameters (above the 1D distribution of the corner plot) when multiple models, or multiple runs, are being loaded in the post-processed. The appropriate tutorial notebook is also provided.
+* The post-processing module has now an option to show the credible intervals of each parameter and run (above the 1D distribution of the corner plot) when multiple runs are plotted in the same figure (but not working for multiple models yet). The appropriate tutorial notebook is also provided.
 
 * Some unit tests and continuous integration.
 
