### 2022-11-10 f154532
Hard upper bound for temperature increased from 7.0 to 7.6
