### 2022-11-11 c32141a
Make version of nbconvert explicit
