+++ .github/workflows/build_docs.yml
@@ -33,7 +33,7 @@ jobs:
         conda install decorator=4.4.1
         pip install sphinxcontrib-websupport==1.1.2
         pip install sphinx_rtd_theme==0.4.3
-        pip install nbconvert=5.6.1=py27_0
+        pip install nbconvert=5.6.1
     - name: Set up GCC
       run: |
          sudo apt-get install -y gcc
