### 2022-11-11 c868502
Maybe installing with conda works
