+++ .github/workflows/build_docs.yml
@@ -30,10 +30,10 @@ jobs:
       run: |
         conda install sphinx=1.8.5
         conda install -c conda-forge nbsphinx=0.5.1
+        conda install -c conda-forge nbconvert=5.6.1
         conda install decorator=4.4.1
         pip install sphinxcontrib-websupport==1.1.2
         pip install sphinx_rtd_theme==0.4.3
-        pip install nbconvert==5.6.1
     - name: Set up GCC
       run: |
          sudo apt-get install -y gcc
