+++ .github/workflows/build_docs.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - main
-  pull_request:
   release:
   workflow_dispatch:
 
