+++ basic_environment.yml
@@ -2,9 +2,9 @@ name: xpsi
 channels:
   - defaults
 dependencies:
-  - python=2.7
-  - numpy=1.16
-  - cython=0.28
-  - matplotlib=2.2
-  - scipy=1.2
-  - wrapt=1.11
+  - python=3.10
+  - numpy
+  - cython
+  - matplotlib
+  - scipy
+  - wrapt
