+++ .github/workflows/ci_tests.yml
@@ -23,7 +23,7 @@ jobs:
       with:
          activate-environment: xpsipy3
          environment-file: basic_environment.yml
-         python-version: 3.10
+         python-version: 3.9
          auto-activate-base: false
     - name: Set up GCC
       run: |
