### 2022-11-11 a6b6db2
remove python requirement
