### 2022-11-14 4323b14
Changelog updated regarding the automated documentation compilation.
