### 2022-11-14 d181cdc
Changelog updated and version changed to v1.1. (#177)
* Changelog updated and version changed to v1.1.

* Changelog updated regarding the automated documentation compilation.
