+++ setup.py
@@ -243,7 +243,7 @@ if __name__ == '__main__':
 
     setup(
         name = 'xpsi',
-        version = '1.0.0',
+        version = '1.1.0',
         author = 'The X-PSI Core Team',
         author_email = 'A.L.Watts@uva.nl',
         url = 'https://github.com/xpsi-group/xpsi',
