+++ xpsi/__init__.py
@@ -4,7 +4,7 @@ An open-source package for neutron star astrostatistics.
 
 """
 from __future__ import print_function
-__version__ = "1.0.0"
+__version__ = "1.1.0"
 __author__ = "The X-PSI Core Team"
 
 try:
