### 2022-11-17 3d160fa
Test compiling docs with h5py installed.
