### 2022-11-17 7b80feb
Previous h5py version not found. Trying without explicit mention of the version.
