+++ .github/workflows/build_docs.yml
@@ -32,7 +32,7 @@ jobs:
         conda install -c conda-forge nbsphinx=0.5.1
         conda install -c conda-forge nbconvert=5.6.1
         conda install decorator=4.4.1
-        conda install h5py=2.10.0=nompi_py27h513d04c_102
+        conda install h5py
         pip install sphinxcontrib-websupport==1.1.2
         pip install sphinx_rtd_theme==0.4.3
     - name: Set up GCC
