### 2022-11-17 e6f9f28
Installing h5py to correctly build up documentation using Github Actions (#181)
* Corrected too short title underlines.

* Test compiling docs with h5py installed (trying with: "conda install h5py=2.10.0=nompi_py27h513d04c_102").

* Previous h5py version not found. Trying without explicit mention of the version.

* Back to building docs only when merging pull requests.
