+++ CHANGELOG.rst
@@ -35,7 +35,7 @@ and this project adheres to
 
 
 [v1.1.0] - 2022-11-14
-~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~
 
 Summary
 ^^^^^^^
@@ -66,7 +66,7 @@ Devarshi Choudhury
 
 
 [v1.0.0] - 2022-09-26
-~~~~~~~~~~~~~~~~~~~
+~~~~~~~~~~~~~~~~~~~~~
 
 Summary
 ^^^^^^^
