+++ .github/workflows/build_docs.yml
@@ -31,6 +31,7 @@ jobs:
         conda install -c conda-forge nbsphinx=0.5.1
         conda install -c conda-forge nbconvert=5.6.1
         conda install decorator=4.4.1
+        conda install h5py
         pip install sphinxcontrib-websupport==1.1.2
         pip install sphinx_rtd_theme==0.4.3
     - name: Set up GCC
