### 2022-11-28 #4977 @JohannesBuchner
Besides the license file link missing, I want to point out that it's not really possible to release a project under MIT license if it depends on GPL libraries. One of your named dependencies is GSL. see https://www.gnu.org/software/gsl/#licensing

Also MultiNest is not open source, so if the project depends on it, you need to add a GPL with exception to allow redistributing as a package. That is what PyMultiNest does as well, so you could adopt the licence from there. You can circumvent this requirement if your project has an interface which allows multiple inference engines (e.g., allow the user to switch in UltraNest instead of MultiNest), and thus is not reliant on running/building with MultiNest.
