### 2022-11-28 #4977 @adonath
Thanks a lot @JohannesBuchner, for pointing out the issue with the license!

@drannawatts I'd propose to resolve this first, before we proceed with a more detailed review. Do you see any issue in re-licensing your project to fulfill the compatibility with the GSL dependency?
