### 2022-11-28 #4977 @matteobachetti
Re-licensing under the GPL potentially limits its future usage (as this exchange shows), so I would be wary of going that path.
Not a lawyer, but I seem to understand that people typically use more permissive licenses (MIT, BSD), so that other people can use them freely and decide whether to release their code under the GPL or not.
If I'm not mistaken, the problem here is just that the authors are *linking* to the GSL in compiled code, right? If it just were an external dependency of a python script, this restriction would not apply?
