### 2022-11-28 #4977 @JohannesBuchner
https://softwareengineering.stackexchange.com/questions/87446/using-a-gplv3-python-module-will-my-entire-project-have-to-be-gplv3-licensed

In any case, cython is compiled here.

An alternative is to rewrite the package without GPL dependencies.
