### 2022-11-28 #4977 @adonath
Thanks @matteobachetti for joining so quickly!

> If I'm not mistaken, the problem here is just that the authors are linking to the GSL in compiled code, right? If it just were an external dependency of a python script, this restriction would not apply?

This is my understanding too.

> An alternative is to rewrite the package without GPL dependencies.

From what I can tell, this seems not a real option now. The GSL routines are used in many, many places for integration and interpolation. This seems like a major refactoring for the project which recently tagged v1.0. However I agree that the code readability and accessibility would actually benefit from relying on e.g. Scipy functionality.

To me it seems the short term option is re-licensing and then including larger refactoring in a roadmap for v2.0.
