### 2022-11-28 #4977 @JohannesBuchner
I agree that re-licensing is the easiest and safest option. Here are the two files you need:

https://github.com/JohannesBuchner/PyMultiNest/blob/master/LICENSE (adjust the first line, and take note of the last paragraph)
and
https://github.com/JohannesBuchner/PyMultiNest/blob/master/COPYING
