### 2022-11-28 #4977 @adonath
I also asked the JOSS editorial team for advice and there was the proposal to maybe choose a dual license here. Here is the quote:

>If they envision possibly replacing GSL, they can say the code is dual licensed (so if you build it with GSL, then you need to comply with GPL, but if you copy out a function that doesn't use GSL, you would only need to comply with a more permissive license).

This would allow complying with GSL, while maximizing the reuse of other parts of the software which might not depend on GSL. The downside is probably just that it's a rare thing to do and might need a more complex license description...
