### 2022-11-28 #4977 @drannawatts
Thank you for all of your input on this!! I think re-licensing is probably the only way to go for now - I'll discuss properly with the rest of the team tomorrow just in case I've missed anything and then get this fixed ASAP. Will update here once it's done!
