### 2022-12-02 f833760
New numerical atmosphere extension added allowing a modified beaming.
