+++ setup.py
@@ -22,6 +22,7 @@ parser = argparse.ArgumentParser(description=desc)
 try:
     parser.add_argument('--NumHot', type=str, help="Numerical atmosphere for the hot region(s)")
     parser.add_argument('--NumElse', type=str, help="Numerical atmosphere for the rest of the surface")
+    parser.add_argument('--NumHotBeam', type=str, help="Numerical atmosphere for the hot region(s) including free beaming")
     # parser.add_argument('--ComptHot', type=str, help="Compton emission model for the hot region(s)")
     # parser.add_argument('--ComptElse', type=str, help="Compton emission model for the rest of the surface")
     if '--help' in sys.argv:
@@ -41,6 +42,10 @@ try:
         print("Copying numerical atmosphere for the rest of the surface")
         shutil.copy('xpsi/surface_radiation_field/archive/elsewhere/numerical.pyx', 'xpsi/surface_radiation_field/elsewhere.pyx')
         sys.argv.remove("--NumElse")
+    if '--NumHotBeam' in sys.argv:
+        print("Copying numerical atmosphere for the hot region(s)")
+        shutil.copy('xpsi/surface_radiation_field/archive/hot/numerical_fbeam.pyx', 'xpsi/surface_radiation_field/hot.pyx')
+        sys.argv.remove("--NumHotBeam")
     # if '--ComptHot' in sys.argv:
     #     print("Copying Compton emission model for the hot region(s)")
     #     shutil.copy('xpsi/surface_radiation_field/archive/hot/compton.pyx', 'xpsi/surface_radiation_field/hot.pyx')
