+++ docs/source/conf.py
@@ -78,9 +78,9 @@ author = u'X-PSI Core Team'
 # built documents.
 #
 # The short X.Y version.
-version = u'1.1'
+version = u'1.2'
 # The full version, including alpha/beta/rc tags.
-release = u'1.1.0'
+release = u'1.2.0'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
