### 2022-12-04 c013d71
New example script added for the beaming extension. Documentation and changelog updated as well.
