+++ examples/examples_modeling_tutorial/TestRun_NumBeam.py
@@ -1,0 +1,611 @@
+'''
+Test script to check that X-PSI installation is working (with the numerical atmosphere extension with free beaming options).
+The run is succesful if it passes the likelihood check (in ~ 1 minute) and sampling finishes without errors (in ~ 5 min).
+The model and synthetic data are based on those shown in the Modelling tutorial notebook, although adding here new parameters on the atmospheric beaming.
+
+Prequisities:
+Before running the script, add the NICER instrument files to the model_data subdirectory:
+nicer_v1.01_arf.txt, nicer_v1.01_rmf_energymap.txt, and nicer_v1.01_rmf_matrix.txt, nsx_H_v200804.out
+(found from https://doi.org/10.5281/zenodo.7094144).
+'''
+
+from __future__ import print_function, division
+
+import os
+import numpy as np
+import math
+import time
+
+from matplotlib import pyplot as plt
+from matplotlib import rcParams
+from matplotlib.ticker import MultipleLocator, AutoLocator, AutoMinorLocator
+from matplotlib import gridspec
+from matplotlib import cm
+
+import xpsi
+
+from xpsi.global_imports import _c, _G, _dpr, gravradius, _csq, _km, _2pi
+
+np.random.seed(xpsi._rank+10)
+
+print('Rank reporting: %d' % xpsi._rank)
+
+#using an example synthetic data
+settings = dict(counts = np.loadtxt('model_data/example_synthetic_realisation.dat', dtype=np.double),
+        channels=np.arange(20,201),
+        phases=np.linspace(0.0, 1.0, 33),
+        first=0, last=180,
+        exposure_time=984307.6661)
+data = xpsi.Data(**settings)
+
+class CustomInstrument(xpsi.Instrument):
+    """ A model of the NICER telescope response. """
+
+    def __call__(self, signal, *args):
+        """ Overwrite base just to show it is possible.
+
+        We loaded only a submatrix of the total instrument response
+        matrix into memory, so here we can simplify the method in the
+        base class.
+
+        """
+        matrix = self.construct_matrix()
+
+        self._folded_signal = np.dot(matrix, signal)
+
+        return self._folded_signal
+
+    @classmethod
+    def from_response_files(cls, ARF, RMF, max_input, min_input=0,
+                            channel_edges=None):
+        """ Constructor which converts response files into :class:`numpy.ndarray`s.
+        :param str ARF: Path to ARF which is compatible with
+                                :func:`numpy.loadtxt`.
+        :param str RMF: Path to RMF which is compatible with
+                                :func:`numpy.loadtxt`.
+        :param str channel_edges: Optional path to edges which is compatible with
+                                  :func:`numpy.loadtxt`.
+        """
+        link ="https://doi.org/10.5281/zenodo.7094144"
+        if min_input != 0:
+            min_input = int(min_input)
+
+        max_input = int(max_input)
+
+        try:
+            ARF = np.loadtxt(ARF, dtype=np.double, skiprows=3)
+        except:
+            print("ERROR: You miss the following file: {}\nThe file is found from here: {}".format(ARF, link))
+            exit()
+
+        try:
+            RMF = np.loadtxt(RMF, dtype=np.double)
+        except:
+            print("ERROR: You miss the following file: {}\nThe file is found from here: {}".format(RMF, link))
+            exit()
+        if channel_edges:
+            try:
+                channel_edges = np.loadtxt(channel_edges, dtype=np.double, skiprows=3)[:,1:]
+            except:
+                print("ERROR: You miss the following file: {}\nThe file is found from here: {}".format(channel_edges, link))
+                exit()
+
+        matrix = np.ascontiguousarray(RMF[min_input:max_input,20:201].T, dtype=np.double)
+
[…]
