+++ examples/examples_modeling_tutorial/modules/CustomHotRegion_Beaming.py
@@ -1,0 +1,234 @@
+from __future__ import division, print_function
+import xpsi
+from xpsi.global_imports import *
+from xpsi import global_imports
+
+from xpsi.cellmesh.mesh_tools import allocate_cells as _allocate_cells
+from xpsi.cellmesh.mesh import construct_spot_cellMesh as _construct_spot_cellMesh
+from xpsi.cellmesh.polar_mesh import construct_polar_cellMesh as _construct_polar_cellMesh
+from xpsi.cellmesh.rays import compute_rays as _compute_rays
+
+from xpsi.Parameter import Parameter, Derive
+from xpsi.ParameterSubspace import ParameterSubspace
+
+
+# Your derived class can have an initializer that sets up your new parameters and the temperature parameters, and then passes them to the base class initializer using the super(<derived class>, self).__init__(*args, **kwargs) call, where kwargs includes custom parameter
+
+class CustomHotRegion(xpsi.HotRegion):
+    """ Implement method for HotRegion with beaming parameters."""
+    required_names = ['super_colatitude',
+                      'super_radius',
+                      'phase_shift',
+                      'super_temperature (if no custom specification)']
+
+    optional_names = ['super_abb',
+                      'super_bbb',
+                      'super_cbb',
+                      'super_dbb',    
+                      'omit_colatitude',
+                      'omit_radius',
+                      'omit_azimuth',
+                      'cede_colatitude',
+                      'cede_radius',
+                      'cede_azimuth',
+                      'cede_temperature']
+
+    def __init__(self,
+                 bounds,
+                 values,
+                 symmetry = True,
+                 omit = False,
+                 cede = False,
+                 concentric = False,
+                 sqrt_num_cells = 32,
+                 min_sqrt_num_cells = 10,
+                 max_sqrt_num_cells = 80,
+                 num_rays = 200,
+                 num_leaves = 64,
+                 num_phases = None,
+                 phases = None,
+                 do_fast = False,
+                 fast_sqrt_num_cells = 16,
+                 fast_min_sqrt_num_cells = 4,
+                 fast_max_sqrt_num_cells = 16,
+                 fast_num_rays = 100,
+                 fast_num_leaves = 32,
+                 fast_num_phases = None,
+                 fast_phases = None,
+                 is_antiphased = False,
+                 custom = None,
+                 image_order_limit = None,
+                 fbeam = False,
+                 **kwargs):
+      
+
+	doc = """
+	log10(superseding region effective temperature [K])
+	"""
+	super_temp = Parameter('super_temperature',
+		  strict_bounds = (3.0, 7.0), # very cold --> very hot
+		  bounds = bounds.get('super_temperature', None),
+		  doc = doc,
+		  symbol = r'$\log_{10}(T\;[\rm{K}])$',
+		  value = values.get('super_temperature', None))
+	if cede:
+		doc = """
+		log10(ceding region effective temperature [K])
+		"""
+		cede_temp = Parameter('cede_temperature',
+			      strict_bounds = (3.0, 7.0), # same story
+			      bounds = bounds.get('cede_temperature', None),
+			      doc = doc,
+			      symbol = r'$\log_{10}(\mathcal{T}\;[\rm{K}])$',
+			      value = values.get('cede_temperature', None))
+	else:
+		cede_temp = None
+
+            
+            
+        if not fbeam: 
+            bounds['super_abb'] = None
+            values['super_abb'] = 0.0
+            no_verb['super_abb'] = True
+            bounds['super_bbb'] = None
+            values['super_bbb'] = 0.0
+            no_verb['super_bbb'] = True
+            bounds['super_cbb'] = None
+            values['super_cbb'] = 0.0
+            no_verb['super_cbb'] = True
+            bounds['super_dbb'] = None
[…]
