### 2022-12-04 90358a8
Specification to the installation comment in Example scripts.
