### 2022-12-09 2e0c52e
Updated licence info and development status in setup.py file
