### 2022-12-09 f80f8e1
Added OSI text back
