### 2022-12-09 8c05297
Updated licence info and development status in setup.py file (#186)
* Updated licence info and development status in setup.py file

* Added OSI text back
