### 2022-12-09 9dc0c11
Conda environment for py3 version renamed.
Also, the fast example run updated to work with py3.
