### 2022-12-09 #4977 @matteobachetti
@drannawatts I tried to install it following the instructions at https://xpsi-group.github.io/xpsi/install.html
When I execute one of the simple example scripts, I get errors like:
```
ERROR:   Could not load MultiNest library "libmultinest.so"
ERROR:   You have to build it first,
ERROR:   and point the LD_LIBRARY_PATH environment variable to it!
ERROR:   manual: https://johannesbuchner.github.io/PyMultiNest/install.html
```
It seems that pymultinest does not find the compiled Multinest library. Maybe there needs to be some addition to `LD_LIBRARY_PATH`?
