### 2022-12-09 #4977 @thjsal
Hi Matteo. The solution is indeed to export the following path:
`export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$HOME/multinest/MultiNest_v3.12_CMake/multinest/lib/`
Seems that this instruction is currently only on the HPC installation page, but it sure should be mentioned on the main installation page as well (we can add it there).
