### 2022-12-09 #4977 @drannawatts
@matteobachetti @JohannesBuchner we are creating GitHub issues tagged 'jossreview' for all of the things that you raise in review.  We'll most likely try to address them all together unless you feel that they are so serious that they are holding you up in which case we'll get them fixed straight away!
