+++ CHANGELOG.rst
@@ -34,7 +34,7 @@ and this project adheres to
 .. ^^^^^^^^^^^
 
 
-[pre-v1.2.1] - 2022-12-10
+[v1.2.1] - 2022-12-11
 ~~~~~~~~~~~~~~~~~~~~~
 
 Summary
