### 2022-12-11 ebd368e
More specific warning explaining why being outside of priors when creating synthetic data.
