### 2022-12-12 fd55a65
Added an option to modify the factor determining when the exact likelihood calculation is skipped. (#191)
