+++ xpsi/likelihoods/default_background_marginalisation.pyx
@@ -131,6 +131,7 @@ def eval_marginal_likelihood(double exposure_time,
                              double sigmas,
                              double llzero,
                              allow_negative = False,
+                             slim = 20.0,
                              background = None):
     """ Evaluate the Poisson likelihood.
 
@@ -221,6 +222,14 @@ def eval_marginal_likelihood(double exposure_time,
         spline from GSL, such oscillations should manifest as small relative
         to those present in cubic splines, for instance, because it is
         designed to handle a rapidly changing second-order derivative.
+
+    :param double slim:
+        The number that determines how many sigmas below the signal from the star
+        can the data be at most (at any channel) before skipping the exact
+        likelihood calculation and returning a random likelihood only slightly
+        above the logZero level of MultiNest (no expected counts or background
+        are returned in that case). By default a value of 20.0 is used.
+        If a negative value is provided, no limit is applied.
 
     :param obj background:
         If not ``None``, then a C-contiguous :class:`numpy.ndarray` of
@@ -382,11 +391,13 @@ def eval_marginal_likelihood(double exposure_time,
             av_STAR += STAR[i,j]
             av_DATA += counts[i,j]
 
-        limit = av_STAR * SCALE - 20.0 * sqrt(av_STAR * SCALE) - av_DATA
-
-        if limit > 0.0:
-            LOGLIKE = llzero * (0.1 + 0.9 * np.random.rand(1))
-            break
+        if slim < 0.0:
+            pass
+        else:
+            limit = av_STAR * SCALE - slim * sqrt(av_STAR * SCALE) - av_DATA
+            if limit > 0.0:
+                LOGLIKE = llzero * (0.1 + 0.9 * np.random.rand(1))
+                break
 
         av_STAR /= n
         av_DATA /= exposure_time
