### 2022-12-12 7b7c533
Changelog and synthetic data example updated.
