### 2022-12-12 4f09d2e
Increased temperature hard bounds and updated warning in synthetize function (#188)
* Hard upper bound for temperature increased from 7.0 to 7.6

* Added text that parameter bounds should be within the atmosphere table

* Warning in synthetize function specified.

* Version number changed, since why not.

* More specific warning explaining why being outside of priors when creating synthetic data.

* Changelog and synthetic data example updated.
