+++ docs/source/conf.py
@@ -80,7 +80,7 @@ author = u'X-PSI Core Team'
 # The short X.Y version.
 version = u'1.2'
 # The full version, including alpha/beta/rc tags.
-release = u'1.2.0'
+release = u'1.2.1'
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
