+++ setup.py
@@ -248,7 +248,7 @@ if __name__ == '__main__':
 
     setup(
         name = 'xpsi',
-        version = '1.2.0',
+        version = '1.2.1',
         author = 'The X-PSI Core Team',
         author_email = 'A.L.Watts@uva.nl',
         url = 'https://github.com/xpsi-group/xpsi',
