+++ xpsi/Everywhere.py
@@ -43,7 +43,9 @@ class Everywhere(ParameterSubspace):
         of a temperature parameter. The parameter name
         ``'temperature'`` must be a key in the dictionary unless the
         parameter is *fixed* or *derived*. If a bound is ``None`` that bound
-        is set equal to a strict hard-coded bound.
+        is set equal to a strict hard-coded bound. We note that the bounds for
+        parameters used in the atmosphere model should be restricted (by the user)
+        to be within the tabulated values, in case a numerical atmosphere extension is used.
 
     :param dict values:
         Either the fixed value of the temperature, a callable if the
@@ -142,7 +144,7 @@ class Everywhere(ParameterSubspace):
 
         if not custom: # setup default temperature parameter
             T = Parameter('temperature',
-                          strict_bounds = (3.0, 7.0), # very cold --> very hot
+                          strict_bounds = (3.0, 7.6), # very cold --> very hot
                           bounds = bounds.get('temperature', None),
                           doc = 'log10(effective temperature [K] everywhere)',
                           symbol = r'$\log_{10}(T\;[\rm{K}])$',
