### 2022-12-14 8757ea9
Replacing kwargs with args to get the triangle plots working with standard getdist.
