### 2022-12-14 866a270
Use weighted_1D_gaussian_kde for KL divergence calculation if (or currently "when") getdist_kde is not possible for nestcheck.
