### 2022-12-14 1b54159
Missing numpy import added for saving credible intervals in corner plotting.
