### 2023-01-06 #4977 @adonath
Happy new year everyone! This is just a quick reminder for @JohannesBuchner and @matteobachetti to continue their review.
Usually JOSS aims for a short review phase between 2-4 weeks, so let's try to aim for concluding the review soon. Please also let me know whether there are any open questions or anything else I can help with. Thanks!
