### 2023-01-08 #4977 @matteobachetti
Hi @adonath , thanks for the reminder. I got a little lost, I was expecting some updates on if the issues got solved, in particular the ones about installation
