### 2023-01-08 #4977 @drannawatts
Hi @matteobachetti, what you were still waiting for from us on installation?  Did the fix that @thjsal gave you on Dec 9 not solve the problem?

We have a github issue raised to make sure that this step (which currently appears only on the HPC installation page) is also copied to the main install instructions, but as I mentioned in my comment on Dec 9 we were thinking to do this 'en masse' with any other issues raised during the review process.  But as I mentioned above we're happy to fix things straight away if they are holding you up!!
