### 2023-01-10 d1114ab
Minor corrections to installation documentation
