### 2023-01-13 #4977 @matteobachetti
@drannawatts the text still lists MIT as the license, I think that needs to be updated. Besides that, I think the paper has all ingredients that JOSS requires.
I installed the code, ran the test, it all seems to work as expected. I've read a couple of papers and attended a couple of seminars that describe the results obtained with a full-blown use of the code, so I have no doubt about the impact of the code or its worthiness of a JOSS paper.
As a small, non-blocking, final comment, I encourage your group to complete the migration to Python 3 (I see an open PR), as most numerical libraries, including your dependencies, have long stopped to make updates and fixes for Python 2.7, and the maintenance of a Python 2.7 environment is becoming more and more unstable.
