### 2023-01-13 #4977 @drannawatts
Oh well spotted on the licence being mentioned in the text of the paper, thanks @matteobachetti ! I will fix that later today.

Python 3 is also imminent -  we actually have a Sprint Week planned for the end of this month to crack the bulk of the remaining work on this transition.  This will then be the X-PSI v2.0 release.
