### 2023-01-16 05737bf
Bug in cross-hair option fixed.
