### 2023-01-16 5b63969
Fixed dpi setting for writing figures when using standard getDist.
