### 2023-01-17 18948ff
deprecated lab_fontsize changed to axes_labelsize
