### 2023-01-25 8acb1dc
Residual plot updated to work with python3 using channel edges.
