+++ xpsi/PostProcessing/_residual.py
@@ -145,8 +145,17 @@ class ResidualPlot(SignalPlot):
         except AttributeError:
             self._set_vminmax()
 
+        #Calculate channel edges by averaging:
+        channels = self._signal.data.channels
+        channel_edges = _np.zeros((len(self._signal.data.channels)+1))
+        channel_edges[1:len(channels)] = (channels[:len(channels)-1]+channels[1:])/2.0
+        chandiff1 = (channels[1]-channels[0])/2.0
+        chandiff2 = (channels[len(channels)-1]-channels[len(channels)-2])/2.0
+        channel_edges[0] = channels[0]-chandiff1
+        channel_edges[len(channels)] = channels[len(channels)-1]+chandiff2
+
         data = self._ax_data.pcolormesh(self._signal.data.phases,
-                                        self._signal.data.channels,
+                                        channel_edges,
                                         self._signal.data.counts/2.0,
                                         cmap = cm.get_cmap(self._data_cmap),
                                         vmin = self._vmin,
@@ -156,7 +165,7 @@ class ResidualPlot(SignalPlot):
         data.set_edgecolor('face')
 
         data = self._ax_data.pcolormesh(self._signal.data.phases + 1.0,
-                                        self._signal.data.channels,
+                                        channel_edges,
                                         self._signal.data.counts/2.0,
                                         cmap = cm.get_cmap(self._data_cmap),
                                         vmin = self._vmin,
@@ -165,8 +174,8 @@ class ResidualPlot(SignalPlot):
                                         rasterized = self._rasterized)
         data.set_edgecolor('face')
 
-        self._ax_data.set_ylim([self._signal.data.channels[0],
-                                self._signal.data.channels[-1]])
+        self._ax_data.set_ylim([channel_edges[0],
+                                channel_edges[-1]])
         self._ax_data.set_yscale('log')
 
         self._data_cb = plt.colorbar(data, cax=self._ax_data_cb,
@@ -184,8 +193,17 @@ class ResidualPlot(SignalPlot):
         except AttributeError:
             self._set_vminmax()
 
+        #Calculate channel edges by averaging:
+        channels = self._signal.data.channels
+        channel_edges = _np.zeros((len(self._signal.data.channels)+1))
+        channel_edges[1:len(channels)] = (channels[:len(channels)-1]+channels[1:])/2.0
+        chandiff1 = (channels[1]-channels[0])/2.0
+        chandiff2 = (channels[len(channels)-1]-channels[len(channels)-2])/2.0
+        channel_edges[0] = channels[0]-chandiff1
+        channel_edges[len(channels)] = channels[len(channels)-1]+chandiff2
+
         model = self._ax_model.pcolormesh(self._signal.data.phases,
-                                          self._signal.data.channels,
+                                          channel_edges,
                                           self.expected_counts/2.0,
                                           cmap = cm.get_cmap(self._model_cmap),
                                           vmin = self._vmin,
@@ -195,7 +213,7 @@ class ResidualPlot(SignalPlot):
         model.set_edgecolor('face')
 
         model = self._ax_model.pcolormesh(self._signal.data.phases + 1.0,
-                                          self._signal.data.channels,
+                                          channel_edges,
                                           self.expected_counts/2.0,
                                           cmap = cm.get_cmap(self._model_cmap),
                                           vmin = self._vmin,
@@ -204,8 +222,8 @@ class ResidualPlot(SignalPlot):
                                           rasterized = self._rasterized)
         model.set_edgecolor('face')
 
-        self._ax_model.set_ylim([self._signal.data.channels[0],
[…]
