### 2023-01-25 505bc34
Option to choose y-scale (lin/log) to residual plot y-axis
