### 2023-01-25 8c00f91
Removed warnings about need for specific getdist and nestcheck versions.
