+++ xpsi/PostProcessing/_global_imports.py
@@ -67,13 +67,6 @@ else:
     if _verbose:
         print('Imported GetDist version: %s' % getdist.__version__)
 
-    _expected_version = '0.3.1'
-    if getdist.__version__ != _expected_version:
-        if _verbose:
-            _warning('The PostProcessing module is compatible with a specific '
-                     'GetDist commit, with version %s, so this module will '
-                     'likely not work as intended.' % _expected_version)
-
     # the following disables getdist.chains.print_load_details
     getdist.chains.print_load_details = False
 
@@ -85,13 +78,6 @@ except ImportError:
 else:
     if _verbose:
         print('Imported nestcheck version: %s' % nestcheck.__version__)
-
-    _expected_version = '0.2.0'
-    if nestcheck.__version__ != _expected_version:
-        if _verbose:
-            _warning('The PostProcessing module is compatible with a specific '
-                     'nestcheck commit, with version %s, so this module will '
-                     'likely not work as intended.' % _expected_version)
 
 class AmbiguityError(xpsiError):
     """ Thrown if ambiguous IDs are declared for objects. """
