### 2023-01-26 dbe3dc5
Commented input parameters used only in the customized nestcheck functions
Changed also the form of fthetas and lims given to bs_param_dists() function
