+++ xpsi/__init__.py
@@ -9,12 +9,10 @@ __author__ = "The X-PSI Core Team"
 
 try:
     __XPSI_SETUP__
-    print("This worked!")
 except NameError:
     __XPSI_SETUP__ = False
 
 if not __XPSI_SETUP__:
-    print("This is not working!")
 
     try:
         from mpi4py import MPI
