### 2023-01-27 02857b9
Indentation problems fixed for py3 in the beaming example modules.
