+++ examples/examples_modeling_tutorial/modules/CustomPrior_Beaming.py
@@ -31,7 +31,7 @@ class CustomPrior(xpsi.Prior):
         #Saving the photoshpere for beaming priors if needed:
         #self.photosphere = photosphere
         
-	super(CustomPrior, self).__init__()
+        super(CustomPrior, self).__init__()
 
     def __call__(self, p = None):
         """ Evaluate distribution at ``p``.
@@ -94,69 +94,69 @@ class CustomPrior(xpsi.Prior):
         abb = self.parameters['p__super_abb']
         bbb = self.parameters['p__super_bbb']
         try:
-		cbb = self.parameters['p__super_cbb']
-		dbb = self.parameters['p__super_dbb']
-	except:
-		cbb = 0.0
-		dbb = 0.0 
+            cbb = self.parameters['p__super_cbb']
+            dbb = self.parameters['p__super_dbb']
+        except:
+            cbb = 0.0
+            dbb = 0.0 
 		
         logg = grav[0] #grav[0] = equatorial radius, grav[1] = polar radius         
 
-	K2keV =  8.61732814974493e-08
-	if (cbb == 0.0 and dbb == 0.0):
-		Elist = np.array([1.0])
-		tempps = [6.0]
-	else:		
-		Elist = np.linspace(-1.5,2.2,num=100) #-1.3,2.0
-		tempps = [self.parameters['p__super_temperature'], self.parameters['s__super_temperature']]
+        K2keV =  8.61732814974493e-08
+        if (cbb == 0.0 and dbb == 0.0):
+            Elist = np.array([1.0])
+            tempps = [6.0]
+        else:		
+            Elist = np.linspace(-1.5,2.2,num=100) #-1.3,2.0
+            tempps = [self.parameters['p__super_temperature'], self.parameters['s__super_temperature']]
 	
-	for isp in range(0,len(tempps)):
-		Ekevp = (10**Elist)*K2keV*(10**tempps[isp])	
-		for ie in range(0,len(Ekevp)):
-			#Check first if beaming function would be negative, giving negative intensities:
-			a = bbb*(Ekevp[ie])**dbb #now named as in f(x)=ax^2+bx+c 
-			b = abb*(Ekevp[ie])**cbb
-			c = 1.0
-			dkr = b**2-4.0*a*c
-			if dkr >= 0:
-			    if abs(a) < 1e-8:
-				if abs(b) > 1e-8:                    
-				    zp_mu = -c/b
-				    if(zp_mu > 0.0 and zp_mu < 1.0):
-					return -np.inf
-			    else:
-				zp1_mu = (-b+np.sqrt(dkr))/(2.0*a)
-				zp2_mu = (-b-np.sqrt(dkr))/(2.0*a)
-				if((zp1_mu > 0.0 and zp1_mu < 1.0) or (zp2_mu > 0.0 and zp2_mu < 1.0)):
-					return -np.inf
+        for isp in range(0,len(tempps)):
+            Ekevp = (10**Elist)*K2keV*(10**tempps[isp])	
+            for ie in range(0,len(Ekevp)):
+                #Check first if beaming function would be negative, giving negative intensities:
+                a = bbb*(Ekevp[ie])**dbb #now named as in f(x)=ax^2+bx+c 
+                b = abb*(Ekevp[ie])**cbb
+                c = 1.0
+                dkr = b**2-4.0*a*c
+                if dkr >= 0:
+                    if abs(a) < 1e-8:
+                        if abs(b) > 1e-8:
+                            zp_mu = -c/b
+                            if(zp_mu > 0.0 and zp_mu < 1.0):
+                                return -np.inf
+                    else:
+                        zp1_mu = (-b+np.sqrt(dkr))/(2.0*a)
[…]
