### 2023-01-27 77a98ec
Added option to fix the random seed for the synthetic data generation with py3.
