### 2023-01-30 c604b31
Changed CI tests to Python 3, test matrix of multiple versions
