### 2023-01-30 cba61e1
Fixed syntax error in ci_tests
