### 2023-01-30 3eb938e
Fixed py2 to py3 conversion issues with raw_input and StringIO
