+++ xpsi/utilities/ProjectionTool.py
@@ -5,6 +5,7 @@ from matplotlib.ticker import NullFormatter
 import sys
 import scipy
 
+from io import StringIO
 import os
 import math
 import time
