+++ docs/source/conf.py
@@ -63,9 +63,29 @@ templates_path = ['_templates']
 # Try to make sidebar with versions:
 html_sidebars = {
     '**': [
-        'versioning.html',
+        'version.html',
     ],
 }
+
+# Whitelist pattern for tags (set to None to ignore all tags)
+smv_tag_whitelist = r'^.*$'
+
+# Whitelist pattern for branches (set to None to ignore all branches)
+smv_branch_whitelist = None
+
+# Whitelist pattern for remotes (set to None to use local branches only)
+smv_remote_whitelist = None
+
+# Pattern for released versions
+smv_released_pattern = r'^tags/.*$'
+
+# Format for versioned output directories inside the build directory
+smv_outputdir_format = '{ref.name}'
+
+# Determines whether remote or local git branches/tags are preferred if their output dirs conflict
+smv_prefer_remote_refs = False
+
+
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
