### 2023-01-30 2cdc05e
Adding explicit creation of html directory
