+++ .github/workflows/build_docs.yml
@@ -21,7 +21,7 @@ jobs:
     - uses: actions/checkout@v3
     - uses: conda-incubator/setup-miniconda@v2
       with:
-         activate-environment: xpsipy2
+         activate-environment: xpsipy3
          environment-file: basic_environment.yml
          python-version: 3.9
          auto-activate-base: false
