+++ .github/workflows/build_docs.yml
@@ -49,7 +49,7 @@ jobs:
       env:
         CC:   gcc-10
     - name : build docs
-      run: sphinx-build docs/ build/
+      run: make html
     - name: Deploy to GitHub Pages
       if: success()
       uses: crazy-max/ghaction-github-pages@v3
