+++ .github/workflows/build_docs.yml
@@ -49,7 +49,8 @@ jobs:
       env:
         CC:   gcc-10
     - name : build docs
-      run: make html
+      run: make
+      working-directory: docs/
     - name: Deploy to GitHub Pages
       if: success()
       uses: crazy-max/ghaction-github-pages@v3
