+++ .github/workflows/build_docs.yml
@@ -49,7 +49,7 @@ jobs:
       env:
         CC:   gcc-10
     - name : build docs
-      run: make versions
+      run: make html
       working-directory: docs/
     - name: Deploy to GitHub Pages
       if: success()
