+++ docs/source/conf.py
@@ -74,7 +74,7 @@ smv_tag_whitelist = r'^.*$'
 smv_branch_whitelist = None
 
 # Whitelist pattern for remotes (set to None to use local branches only)
-smv_remote_whitelist = None
+smv_remote_whitelist = r'^(origin)$'
 
 # Pattern for released versions
 smv_released_pattern = r'^tags/.*$'
