### 2023-01-30 65bad14
Changed configuration pattern for multiversion
