+++ docs/source/conf.py
@@ -69,7 +69,8 @@ html_sidebars = {
 
 
 smv_rebuild_tags = False
-smv_branch_whitelist = r'py3_multiversion_docs'
+#smv_branch_whitelist = r'py3_multiversion_docs'
+smv_branch_whitelist = r'^(?!py3_multiversion_docs).*$'
 smv_released_pattern = r'^tags/v.*$'
 
 smv_tag_whitelist = r'^.*$'
