+++ docs/source/conf.py
@@ -68,20 +68,19 @@ html_sidebars = {
 }
 
 
-smv_rebuild_tags = False
-smv_branch_whitelist = r'py3_multiversion_docs'
+#smv_rebuild_tags = False
+#smv_branch_whitelist = r'py3_multiversion_docs'
+
+#smv_remote_whitelist = r'^origin$'
+smv_prefer_remote_refs = True
+
+### OTHERS
 #smv_released_pattern = r'^tags/v.*$'
-smv_released_pattern = "None"
-
 #smv_tag_whitelist = r'^.*$'
-smv_tag_whitelist = "None"
-
 #smv_branch_whitelist = r'^.*$'
 #smv_branch_whitelist = r"^main$"  # build main branch, and anything relating to documentation
-
 #smv_released_pattern = r'.*tags.*'
-smv_remote_whitelist = r'^origin$'
-smv_prefer_remote_refs = True
+
 
 
 # The suffix(es) of source filenames.
