### 2023-01-30 40c1b88
updated install page to python 3 up to not including multinest
