+++ docs/source/install.rst
@@ -10,181 +10,104 @@ X-PSI is an open-source software package that is available on `GitHub
 
     git clone https://github.com/xpsi-group/xpsi.git </path/to/xpsi>
 
-In this page, we lay down the instructions for installing X-PSI and all the  necessary pre-requisites on your local self-administered system.
+In this page, we lay down the instructions for installing X-PSI and all the
+necessary pre-requisites on your local self-administered system.
 
 .. note::
 
-    For installation on a high-performance computing system, we direct the reader to the :ref:`hpcsystems` page for guidance since the instructions on this page are either not applicable or do not target performance.
+    For installation on a high-performance computing system, we direct the 
+    reader to the :ref:`hpcsystems` page for guidance since the instructions 
+    on this page are either not applicable or do not target performance.
 
 .. _dev_env:
 
 Python environment
 ------------------
 
-X-PSI was developed in Python 2.7, and is in the process of being ported to Python 3.
-Fortunately, there are several ways to create a virtual environment with a
-different version of Python, without disrupting your Python ecosystem.
-
-This section is divided into two subsections. We recommend that the user follow the instructions in the :ref:`basic_env` subsection to begin with. If faced with installation issues, the user may refer to the :ref:`diagnosis_env` subsection.
-
-
+X-PSI was originally developed in Python 2.7 and was ported to Python 3 as of 
+X-PSI 2. We recommend creating a conda virtual environment with anaconda 3 as
+per instructions below so as to not disrupt your Python ecosystem.
 
 .. _basic_env:
 
-Basic Conda environment
-^^^^^^^^^^^^^^^^^^^^^^^
+Conda Environment
+^^^^^^^^^^^^^^^^^
 
-In the source directory we provide a basic dependency file that installs
-the core Python packages required for *likelihood* functionality. These
-packages are:
+In the source directory we provide a dependency file that installs
+the Python packages required for full functionality of X-PSI:
+
+.. code-block:: 
+    name: xpsi2
+    channels:
+      - defaults
+      - conda-forge
+    dependencies:
+      - numpy
+      - cython
+      - matplotlib
+      - scipy
+      - wrapt
+      - pymultinest  # nested sampling
+      - mpi4py  # nested sampling
+      - getdist  # posterior KDE corner plotting
+      - h5py  # storage of X-ray signals computed from posterior samples
+      - nestcheck  # posterior error analysis, plotting, run combination, etc.
+      - fgivenx  # conditional posterior plotting; also required by nestcheck
+
+The core packages required for likelihood functionality are:
 
 * `NumPy <https://docs.scipy.org/doc/numpy/index.html>`_
 * `Cython <http://cython.readthedocs.io/en/latest>`_
+* `Matplotlib <https://matplotlib.org/stable/index.html>`_
+* `Scipy <https://docs.scipy.org/doc//scipy/index.html>`_
+* `Wrapt <https://wrapt.readthedocs.io/en/latest/>`_
+
+Then, optional packages required for nested sampling are: 
+
+* `pyMultiNest <https://johannesbuchner.github.io/PyMultiNest/>`_
+* `mpi4py <http://cython.readthedocs.io/en/latest>`_
+
+Note that pyMultiNest expects a MultiNest, which can be installed 
+afterwards. Installation instructions for `multinest`__ are given below. 
+Finally, optional packages required for post-processing are:
+
+* `getdist <https://getdist.readthedocs.io/en/latest/>`_
+* `h5py <https://docs.h5py.org/en/stable/index.html>`_
+* `nestcheck <https://nestcheck.readthedocs.io/en/latest/>`_
+* `fgivenx <https://fgivenx.readthedocs.io/en/latest/>`_
 
[…]
