+++ environment.yml
@@ -1,173 +1,17 @@
-name: xpsi
+name: xpsitest
 channels:
-  - conda-forge
-  - defaults
+        - defaults
+        - conda-forge
 dependencies:
-  - _libgcc_mutex=0.1=main
-  - alabaster=0.7.12=py27_0
-  - algopy=0.5.7=py_0
-  - asn1crypto=1.4.0=py_0
-  - atomicwrites=1.4.0=py_0
-  - attrs=20.1.0=py_0
-  - babel=2.8.0=py_0
-  - backports=1.0=py_2
-  - backports.functools_lru_cache=1.6.1=py_0
-  - backports.shutil_get_terminal_size=1.0.0=py27_2
-  - backports_abc=0.5=py_1
-  - blas=1.0=mkl
-  - bleach=3.1.5=py_0
-  - bzip2=1.0.8=h7b6447c_0
-  - ca-certificates=2020.7.22=0
-  - certifi=2019.11.28=py27_0
-  - cffi=1.13.2=py27h2e261b9_0
-  - chardet=3.0.4=py27_1003
-  - configparser=4.0.2=py27_0
-  - contextlib2=0.6.0.post1=py_0
-  - cryptography=2.3.1=py27hc365091_0
-  - cycler=0.10.0=py27_0
-  - cython=0.28.5=py27hf484d3e_0
-  - dbus=1.13.16=hb2f20db_0
-  - decorator=4.4.2=py_0
-  - defusedxml=0.6.0=py_0
-  - docutils=0.15.2=py27_0
-  - entrypoints=0.3=py27_0
-  - enum34=1.1.6=py27_1
-  - expat=2.2.9=he6710b0_2
-  - ffmpeg=4.2.2=h20bf706_0
-  - fontconfig=2.13.0=h9420a91_0
-  - freetype=2.10.2=h5ab3b9f_0
-  - funcsigs=1.0.2=py27_0
-  - functools32=3.2.3.2=py27_1
-  - futures=3.3.0=py27_0
-  - glib=2.63.1=h5a9c865_0
-  - gmp=6.1.2=h6c8ec71_1
-  - gnutls=3.6.5=h71b1129_1002
-  - gst-plugins-base=1.14.0=hbbd80ab_1
-  - gstreamer=1.14.0=hb453b48_1
-  - h5py=2.9.0=py27h7918eee_0
-  - hdf5=1.10.4=hb1b8bf9_0
-  - icu=58.2=he6710b0_3
-  - idna=2.10=py_0
-  - imagesize=1.2.0=py_0
-  - importlib_metadata=1.3.0=py27_0
-  - intel-openmp=2019.4=243
-  - ipaddress=1.0.23=py_0
-  - ipykernel=4.10.0=py27_0
-  - ipython=5.8.0=py27_0
-  - ipython_genutils=0.2.0=py27_0
-  - ipywidgets=7.5.1=py_0
-  - jinja2=2.11.2=py_0
-  - joblib=0.14.1=py_0
-  - jpeg=9b=h024ee3a_2
-  - jsonschema=3.2.0=py27_0
-  - jupyter=1.0.0=py27_7
-  - jupyter_client=5.3.4=py27_0
-  - jupyter_console=5.2.0=py27_1
-  - jupyter_core=4.6.1=py27_0
-  - kiwisolver=1.1.0=py27he6710b0_0
-  - lame=3.100=h7b6447c_0
-  - libedit=3.1=heed3624_0
-  - libffi=3.2.1=hd88cf55_4
-  - libgcc-ng=9.1.0=hdf63c60_0
-  - libgfortran-ng=7.3.0=hdf63c60_0
-  - libopus=1.3.1=h7b6447c_0
-  - libpng=1.6.37=hbc83047_0
[…]
