### 2023-01-30 e987909
Not anymore used customized codes and parameters for bootstrapping removed.
