+++ xpsi/PostProcessing/_corner.py
@@ -208,8 +208,6 @@ class CornerPlotter(PostProcessor):
     def _plot_density_with_error(self,
                                  plotter = None,
                                  fthetas = None,
-                                 kde_func = None,
-                                 kde_kwargs = None,
                                  **kwargs):
         """
         :param plotter:
@@ -223,14 +221,6 @@ class CornerPlotter(PostProcessor):
             functions. Additional functions are always plotted using the
             native :mod:`nestcheck` matplotlib figure; the parameter densities
             are be added to a :mod:`getdist` lower-triangle plot is supplied.
-
-        :param func kde_func:
-            Function for KDE compatible with :mod:`nestcheck.plots`. Must
-            be *weighted* KDE (Higson et al. 2018). If ``None``, uses
-            :mod:`getdist` if available, or the native KDE function otherwise.
-            If using :mod:`getdist`, the KDE settings are automatically
-            retrieved from the first run and applied to :mod:`nestcheck` and
-            :mod:`fgivenx` *for all runs*.
 
         :param kwargs:
             Keyword arguments for :func:`nestcheck.plots.bs_param_dists`.
@@ -267,23 +257,6 @@ class CornerPlotter(PostProcessor):
                 l[0] = (l[0] if l[0] > b[0] else b[0])
                 l[1] = (l[1] if l[1] < b[1] else b[1])
 
-        if kde_func is None:
-            try:
-                kde_func = getdist_kde
-            except NameError:
-                kde_func = weighted_1d_gaussian_kde
-                kde_kwargs = [None] * len(runs)
-            else:
-                normalize = kwargs.pop('normalize', False)
-                kde_kwargs = []
-                for run in runs:
-                    kde_kwargs.append(
-                            {'settings': run.kde_settings,
-                             'ranges': [run.bounds[param] for param in params],
-                             'normalize': normalize}
-                            )
-
-        lines = kwargs.pop('lines', False)
         parallel = kwargs.pop('parallel', True)
         rasterize_contours = kwargs.pop('rasterize_contours', True)
         tqdm_kwargs = kwargs.pop('tqdm_kwargs', None)
@@ -297,20 +270,14 @@ class CornerPlotter(PostProcessor):
                      'Added density error information'):
             fig = bs_param_dists(nestcheck_bcknds,
                                  fthetas=_fthetas[0],
-                                 #kde_func=kde_func,
-                                 #kde_kwargs=kde_kwargs,
                                  ftheta_lims=lims[0],
                                  nx=nx,
                                  ny=ny,
                                  n_simulate=n_simulate,
-                                 #simulate_weights=True,
-                                 #getdist_plotter=plotter,
                                  figsize=figsize,
-                                 #lines=lines,
                                  parallel=parallel,
                                  rasterize_contours=rasterize_contours,
                                  labels=labels,
-                                 #no_means=True,
                                  tqdm_kwargs=tqdm_kwargs)
 
         if fig: figs.append(fig)
@@ -320,7 +287,6 @@ class CornerPlotter(PostProcessor):
                 num_funcs = len(fthetas[0])
             else:
                 num_funcs = len(fthetas)
-            kde_kwargs['ranges'] = ftheta_lims
             figsize *= float(num_funcs)/len(params)
             if 'ftheta_labels' in kwargs:
                 kwargs = {'labels': kwargs['ftheta_labels']}
@@ -329,16 +295,11 @@ class CornerPlotter(PostProcessor):
 
             fig = bs_param_dists(nestcheck_bcknds,
                                  fthetas=fthetas,
-                                 #kde_func=kde_func,
-                                 #kde_kwargs=kde_kwargs,
                                  ftheta_lims=ftheta_lims,
                                  nx=nx,
[…]
