+++ docs/Makefile
@@ -2,7 +2,7 @@
 #
 
 # You can set these variables from the command line.
-SPHINXOPTS    =
+SPHINXOPTS    = --dump-metadata
 SPHINXBUILD   = sphinx-build
 SPHINXMULTI   = sphinx-multiversion
 SPHINXPROJ    = xpsi
