### 2023-01-31 da7d954
Reduced marker size, removed grid, added beginning of model check
