### 2023-01-31 1a0aea9
Updating X-PSI py3 for standard getdist and nestcheck. (#192)
* Switching to standard NestCheck namings.

* Replacing kwargs with args to get the triangle plots working with standard getdist.

* Use weighted_1D_gaussian_kde for KL divergence calculation if (or currently "when") getdist_kde is not possible for nestcheck.

* Missing numpy import added for saving credible intervals in corner plotting.

* Bug in cross-hair option fixed.

* Fixed dpi setting for writing figures when using standard getDist.

* deprecated lab_fontsize changed to axes_labelsize

* Residual plot updated to work with py
[…]
