+++ xpsi/PostProcessing/_residual.py
@@ -82,6 +82,11 @@ class ResidualPlot(SignalPlot):
             ax.xaxis.set_minor_locator(MultipleLocator(0.05))
             ax.set_xlim([0.0,2.0])
 
+        if "yscale" in kwargs:
+            self.yscale = kwargs.get("yscale")
+        else:
+            self.yscale = "log"
+
         plt.close()
 
     @make_verbose('ResidualPlot object iterating over samples',
@@ -145,8 +150,17 @@ class ResidualPlot(SignalPlot):
         except AttributeError:
             self._set_vminmax()
 
+        #Calculate channel edges by averaging:
+        channels = self._signal.data.channels
+        channel_edges = _np.zeros((len(self._signal.data.channels)+1))
+        channel_edges[1:len(channels)] = (channels[:len(channels)-1]+channels[1:])/2.0
+        chandiff1 = (channels[1]-channels[0])/2.0
+        chandiff2 = (channels[len(channels)-1]-channels[len(channels)-2])/2.0
+        channel_edges[0] = channels[0]-chandiff1
+        channel_edges[len(channels)] = channels[len(channels)-1]+chandiff2
+
         data = self._ax_data.pcolormesh(self._signal.data.phases,
-                                        self._signal.data.channels,
+                                        channel_edges,
                                         self._signal.data.counts/2.0,
                                         cmap = cm.get_cmap(self._data_cmap),
                                         vmin = self._vmin,
@@ -156,7 +170,7 @@ class ResidualPlot(SignalPlot):
         data.set_edgecolor('face')
 
         data = self._ax_data.pcolormesh(self._signal.data.phases + 1.0,
-                                        self._signal.data.channels,
+                                        channel_edges,
                                         self._signal.data.counts/2.0,
                                         cmap = cm.get_cmap(self._data_cmap),
                                         vmin = self._vmin,
@@ -165,9 +179,9 @@ class ResidualPlot(SignalPlot):
                                         rasterized = self._rasterized)
         data.set_edgecolor('face')
 
-        self._ax_data.set_ylim([self._signal.data.channels[0],
-                                self._signal.data.channels[-1]])
-        self._ax_data.set_yscale('log')
+        self._ax_data.set_ylim([channel_edges[0],
+                                channel_edges[-1]])
+        self._ax_data.set_yscale(self.yscale)
 
         self._data_cb = plt.colorbar(data, cax=self._ax_data_cb,
                                      ticks=_get_default_locator(None),
@@ -184,8 +198,17 @@ class ResidualPlot(SignalPlot):
         except AttributeError:
             self._set_vminmax()
 
+        #Calculate channel edges by averaging:
+        channels = self._signal.data.channels
+        channel_edges = _np.zeros((len(self._signal.data.channels)+1))
+        channel_edges[1:len(channels)] = (channels[:len(channels)-1]+channels[1:])/2.0
+        chandiff1 = (channels[1]-channels[0])/2.0
+        chandiff2 = (channels[len(channels)-1]-channels[len(channels)-2])/2.0
+        channel_edges[0] = channels[0]-chandiff1
+        channel_edges[len(channels)] = channels[len(channels)-1]+chandiff2
+
         model = self._ax_model.pcolormesh(self._signal.data.phases,
-                                          self._signal.data.channels,
+                                          channel_edges,
                                           self.expected_counts/2.0,
                                           cmap = cm.get_cmap(self._model_cmap),
                                           vmin = self._vmin,
@@ -195,7 +218,7 @@ class ResidualPlot(SignalPlot):
         model.set_edgecolor('face')
 
         model = self._ax_model.pcolormesh(self._signal.data.phases + 1.0,
-                                          self._signal.data.channels,
[…]
