### 2023-01-31 3be13fe
Checks on models done, single hot spot still does not work
