+++ xpsi/utilities/ProjectionTool.py
@@ -126,7 +126,7 @@ def transform(thetaR,phiR,V,phi0=0.0):
 
 def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase = True, SaveFlag = False, dir = "", Name = "", POVname="", extension = ".png", antipodal = False):
     """
-    GOAL: printing 2D visualisation of hot spot
+    GOAL: printing 2D visualisation of hot spot(s) on a spherical surface
     ##### INPUTS #####
     Vp: vector of parameters
     model: model adopted e.g. ST-U, ST-PST !PLEASE USE "-S" and not just "S" for symmetric models
@@ -165,30 +165,99 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
     except AttributeError:
         DICT_VECTOR = dictVp.names
         
+    def hotspot_check(hot_name, params, symbol_type_hot, asymFlag = False):
+        """
+        GOAL: check consistency between hot spot name and params
+        ##### INPUTS #####
+        hot_name: hot spot names according to our naming convention (P,E,C)ST/(P,E,C)DT
+        params: list of parameters
+        symbol_type_hot: "" empty for single hot spot; "p" for primary hot spot; "s" for secondary hot spot
+        asymFlag: if True it allows a "partially derived" hot spot to be described only by temperature and radius
+        ##### OUTPUTS ####
+        -
+        """
+        ### check super
+        T_name = symbol_type_hot+'__super_temperature'
+        R_name = symbol_type_hot+'__super_radius'
+        P_name = symbol_type_hot+'__phase_shift'
+        C_name = symbol_type_hot+'__super_colatitude'
+        if ((not(R_name in params) or not(T_name in params))) or (not(asymFlag) and (not(P_name in params) or not(C_name in params))):
+            #print ("here:",((not(R_name in params) or not(T_name in params))))
+            #print ("here:",(not(asymFlag) and (not(P_name in params) or not(C_name in params))))
+            #if ((not(T_name in params) and not(R_name in params))) or (not(asymFlag) and (not(P_name in params) or not(C_name in params))):
+            print ("ERROR! super properties required for \'%s\' hot spot ('p' = primary; 's' = secondary and '' for single hot spot model)"%symbol_type_hot)
+            raise IpyExit
+        elif (asymFlag and ((P_name in params) or (C_name in params))):
+            print ("WARNING! there are info for a secondary hot spot that will not being used")
+        
+        if ('DT' in hot_name):
+            ### check cede
+            T_name = symbol_type_hot+'__cede_temperature'
+            R_name = symbol_type_hot+'__cede_radius'
+            if ('C' in hot_name) and not((T_name in params) or (R_name in params)):
+                print ("ERROR! Double temperature (DT) models require the definition of cede properties")
+                raise IpyExit
+            C_name = symbol_type_hot+'__cede_colatitude'
+            A_name = symbol_type_hot+'__cede_azimuth'
+            
+            if ('C' in hot_name) and ((C_name in params) or (A_name in params)):
+                print ("WARNING! there are info for a complex geometry, but they are not being used")
+            if (('P' in hot_name) or ('E' in hot_name)) and not((C_name in params) or (C_name in params)):
+                print ("ERROR! Protruding and Eccentric models require the definition of cede colatitudes and phases")
+                raise IpyExit
+            if not(('C' in hot_name) ^ ('E' in hot_name) ^ ('P 'in hot_name)):
+                print ("ERROR! Double temperature models require the setting if concentric (C, hot spot name = CDT), eccentric (E, hot spot name = EDT), protruding (P, hot spot name = PDT)")
[…]
