### 2023-01-31 a803417
Single hot spot works, checks of model done - missing warning for not using
