+++ xpsi/utilities/ProjectionTool.py
@@ -97,6 +97,9 @@ ABBl = ['PS','SC','SR','ST','OA','OC','OR','CA','CC','CR','CT']
 REF = dict(zip(ABBl,REFl))
 
 def transform(thetaR,phiR,V,phi0=0.0):
+        """
+        
+        """
         phi0 = phi0*np.pi*2.
 
         RA = [-np.sin(phi0),np.cos(phi0),0.]
@@ -129,7 +132,7 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
     GOAL: printing 2D visualisation of hot spot(s) on a spherical surface
     ##### INPUTS #####
     Vp: vector of parameters
-    model: model adopted e.g. ST-U, ST-PST !PLEASE USE "-S" and not just "S" for symmetric models
+    model: model adopted e.g. ST-U, ST+PST !PLEASE USE "-S" and not just "S" for symmetric models. Recognised hot spot names are "ST", "PST", "CST", "EST", "PDT", "CDT", "EDT"; for two hot spot model connect each hot spot model with "+" (first for primary; second for secondary). If the two hot spots have the same name, add "-S" (if all the secondary hot spot is antipodal and its temperature and radius is the same of the primary); "-U" if the properties of the secondary spot is completely unrelated to the primary ones; "-Ua" if the secondary hot spot is antipodal to the primary but has independent "radius" and "temperature".
     POV: Point Of View - location (for string arguments: colatitude) from which the neutron star is observed; if string can be:
         "I" -from the point of view of Earth-: in this case x and y are rotated compared
                                            other configuration to visualise equator horizontally and North on top
@@ -177,28 +180,26 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
         -
         """
         ### check super
-        T_name = symbol_type_hot+'__super_temperature'
-        R_name = symbol_type_hot+'__super_radius'
-        P_name = symbol_type_hot+'__phase_shift'
-        C_name = symbol_type_hot+'__super_colatitude'
+        T_name = symbol_type_hot+REF['ST']
+        R_name = symbol_type_hot+REF['SR']#'__super_radius'
+        P_name = symbol_type_hot+REF['PS']#'__phase_shift'
+        C_name = symbol_type_hot+REF['SC']#'__super_colatitude'
         if ((not(R_name in params) or not(T_name in params))) or (not(asymFlag) and (not(P_name in params) or not(C_name in params))):
-            #print ("here:",((not(R_name in params) or not(T_name in params))))
-            #print ("here:",(not(asymFlag) and (not(P_name in params) or not(C_name in params))))
-            #if ((not(T_name in params) and not(R_name in params))) or (not(asymFlag) and (not(P_name in params) or not(C_name in params))):
             print ("ERROR! super properties required for \'%s\' hot spot ('p' = primary; 's' = secondary and '' for single hot spot model)"%symbol_type_hot)
             raise IpyExit
         elif (asymFlag and ((P_name in params) or (C_name in params))):
             print ("WARNING! there are info for a secondary hot spot that will not being used")
+            
         
         if ('DT' in hot_name):
             ### check cede
-            T_name = symbol_type_hot+'__cede_temperature'
-            R_name = symbol_type_hot+'__cede_radius'
+            T_name = symbol_type_hot+REF['CT']#'__cede_temperature'
+            R_name = symbol_type_hot+REF['CR']
[…]
