### 2023-01-31 e658af2
Updated `module_generator` docstrings
