### 2023-01-31 54a1a9a
Added -a option to Makefile for sphinxmulti
