+++ .github/workflows/build_docs.yml
@@ -57,7 +57,7 @@ jobs:
       uses: crazy-max/ghaction-github-pages@v3
       with:
         target_branch: gh-pages
-        build_dir: docs/build/html
+        build_dir: docs/build/
         jekyll: false
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
