+++ xpsi/utilities/ProjectionTool.py
@@ -151,6 +151,7 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
     returns ax (ax = fig.add_subplot(111))
     """
 
+    # Set abbreviations for plot labels
     Plab = 'Primary'
     Slab = 'Secondary'
     EMIlab = ' emission'
@@ -162,30 +163,31 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
     nColors   = 7
     mycolors0 = [cm(xcol) for xcol in np.linspace(0,1 , nColors)]
 
+    # Define set of parameters
     DICT_VECTOR = []
     try:
         DICT_VECTOR = dictVp.keys()
     except AttributeError:
         DICT_VECTOR = dictVp.names
         
-    def hotspot_check(hot_name, params, symbol_type_hot, asymFlag = False):
+    def hotspot_check(hot_name, params, HStag, asymFlag = False):
         """
         GOAL: check consistency between hot spot name and params
         ##### INPUTS #####
         hot_name: hot spot names according to our naming convention (P,E,C)ST/(P,E,C)DT
         params: list of parameters
-        symbol_type_hot: "" empty for single hot spot; "p" for primary hot spot; "s" for secondary hot spot
+        HStag: Hot Spot tag, "p" for primary hot spot (also for single hot spot models); "s" for secondary hot spot
         asymFlag: if True it allows a "partially derived" hot spot to be described only by temperature and radius
         ##### OUTPUTS ####
         -
         """
         ### check super
-        T_name = symbol_type_hot+REF['ST']
-        R_name = symbol_type_hot+REF['SR']#'__super_radius'
-        P_name = symbol_type_hot+REF['PS']#'__phase_shift'
-        C_name = symbol_type_hot+REF['SC']#'__super_colatitude'
+        T_name = HStag+REF['ST']
+        R_name = HStag+REF['SR']#'__super_radius'
+        P_name = HStag+REF['PS']#'__phase_shift'
+        C_name = HStag+REF['SC']#'__super_colatitude'
         if ((not(R_name in params) or not(T_name in params))) or (not(asymFlag) and (not(P_name in params) or not(C_name in params))):
-            print ("ERROR! super properties required for \'%s\' hot spot ('p' = primary; 's' = secondary and '' for single hot spot model)"%symbol_type_hot)
+            print ("ERROR! super properties required for \'%s\' hot spot ('p' = primary; 's' = secondary and '' for single hot spot model)"%HStag)
             raise IpyExit
         elif (asymFlag and ((P_name in params) or (C_name in params))):
             print ("WARNING! there are info for a secondary hot spot that will not being used")
@@ -193,41 +195,52 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
         
         if ('DT' in hot_name):
             ### check cede
-            T_name = symbol_type_hot+REF['CT']#'__cede_temperature'
-            R_name = symbol_type_hot+REF['CR']
+            T_name = HStag+REF['CT']
+            R_name = HStag+REF['CR']
             if ('C' in hot_name) and not((T_name in params) or (R_name in params)):
                 print ("ERROR! Double temperature (DT) models require the definition of cede properties")
                 raise IpyExit
-            C_name = symbol_type_hot+REF['CT']#'__cede_colatitude'
-            A_name = symbol_type_hot+REF['CA']#'__cede_azimuth'
+            C_name = HStag+REF['CT']
+            A_name = HStag+REF['CA']
             
[…]
