### 2023-01-31 #4977 @JohannesBuchner
Hi all,

I am alive.

Also made some progress today.

I am struggling to get jupyter notebooks to run on py2/conda, it stops with a 500 error at
```
File "/home/user/miniconda3/envs/xpsi/lib/python2.7/site-packages/notebook/notebook/handlers.py", line 19, in get_custom_frontend_exporters
        from nbconvert.exporters.base import get_export_names, get_exporter
    ImportError: No module named base
```
I also do not manage to install getdist on conda with python=2.7.

But I managed to run examples/examples_modeling_tutorial, at least to where it runs multinest, which I count as functional success.

CI tests look good, and very nice documentation!

You could keep track of your test coverage in the CI and submit it to coveralls (or whichever is your favorite) and add a button.

You could accelerate your CI here https://github.com/xpsi-group/xpsi/blob/main/.github/workflows/build_docs.yml#L30 and save some CO2 by putting all the conda installs into one line to avoid rebuilding the dependency tree. Same for the pip commands. Mamba is a popular faster alternative conda interface you could also look into.

This all looks quite sensible, so I am recommending acceptance.
