### 2023-02-01 81a7009
Moving conda install commands to one line to speed up build
