+++ .github/workflows/build_docs.yml
@@ -27,13 +27,8 @@ jobs:
          auto-activate-base: false
     - name : install docs dependencies
       run: |
-        conda install sphinx=1.8.5
-        conda install -c conda-forge nbsphinx=0.5.1
-        conda install -c conda-forge nbconvert=5.6.1
-        conda install decorator=4.4.1
-        conda install h5py
-        pip install sphinxcontrib-websupport==1.1.2
-        pip install sphinx_rtd_theme==0.4.3
+        conda install sphinx=1.8.5 decorator=4.4.1 h5py -c conda-forge nbsphinx=0.5.1 nbconvert=5.6.1 
+        pip install sphinxcontrib-websupport==1.1.2 sphinx_rtd_theme==0.4.3
     - name: Set up GCC
       run: |
          sudo apt-get install -y gcc
