+++ .github/workflows/build_docs.yml
@@ -27,7 +27,8 @@ jobs:
          auto-activate-base: false
     - name : install docs dependencies
       run: |
-        conda install sphinx=1.8.5 decorator=4.4.1 h5py -c conda-forge nbsphinx=0.5.1 nbconvert=5.6.1 
+        conda install sphinx=1.8.5 decorator=4.4.1 h5py 
+        conda install -c conda-forge nbsphinx=0.5.1 nbconvert=5.6.1
         pip install sphinxcontrib-websupport==1.1.2 sphinx_rtd_theme==0.4.3
     - name: Set up GCC
       run: |
