+++ .github/workflows/build_docs.yml
@@ -19,12 +19,20 @@ jobs:
 
     steps:
     - uses: actions/checkout@v3
-    - uses: conda-incubator/setup-miniconda@v2
+
+    - name: Install Conda environment with Micromamba
+      uses: mamba-org/provision-with-micromamba@main
       with:
-         activate-environment: xpsipy2
-         environment-file: basic_environment.yml
-         python-version: 2.7
-         auto-activate-base: false
+        environment-file: basic_environment.yml
+        environment-name: xpsipy2
+        extra-specs: |
+          python=2.7
+#    - uses: conda-incubator/setup-miniconda@v2
+#      with:
+#         activate-environment: xpsipy2
+#         environment-file: basic_environment.yml
+#         python-version: 2.7
+#         auto-activate-base: false
     - name : install docs dependencies
       run: |
         conda install sphinx=1.8.5 decorator=4.4.1 h5py 
