+++ .github/workflows/build_docs.yml
@@ -23,20 +23,16 @@ jobs:
     - name: Install Conda environment with Micromamba
       uses: mamba-org/provision-with-micromamba@main
       with:
+        cache-env: true
+        cache-env-key: env-envkey-${{ github.sha }}-${{ github.run_attempt }}
         environment-file: basic_environment.yml
         environment-name: xpsipy2
         extra-specs: |
           python=2.7
-#    - uses: conda-incubator/setup-miniconda@v2
-#      with:
-#         activate-environment: xpsipy2
-#         environment-file: basic_environment.yml
-#         python-version: 2.7
-#         auto-activate-base: false
     - name : install docs dependencies
       run: |
-        conda install sphinx=1.8.5 decorator=4.4.1 h5py 
-        conda install -c conda-forge nbsphinx=0.5.1 nbconvert=5.6.1
+        mamba install sphinx=1.8.5 decorator=4.4.1 h5py 
+        mamba install -c conda-forge nbsphinx=0.5.1 nbconvert=5.6.1
         pip install sphinxcontrib-websupport==1.1.2 sphinx_rtd_theme==0.4.3
     - name: Set up GCC
       run: |
