+++ .github/workflows/build_docs.yml
@@ -31,8 +31,8 @@ jobs:
           python=2.7
     - name : install docs dependencies
       run: |
-        mamba install sphinx=1.8.5 decorator=4.4.1 h5py 
-        mamba install -c conda-forge nbsphinx=0.5.1 nbconvert=5.6.1
+        micromamba install sphinx=1.8.5 decorator=4.4.1 h5py 
+        micromamba install -c conda-forge nbsphinx=0.5.1 nbconvert=5.6.1
         pip install sphinxcontrib-websupport==1.1.2 sphinx_rtd_theme==0.4.3
     - name: Set up GCC
       run: |
