### 2023-02-01 d09e595
Mandatory channel_edges
In all examples and notebook, the channel_edges was an optional argument of the `CustomInstrument`, but the return of the `from_response_files()` function would crash if no channel_edges was provided (since it is returning `channel_edges[..,..]`.
So now, `channel_edges` is a mandatory argument.
