### 2023-02-01 #4977 @dhuppenkothen
@JohannesBuchner Thanks for the comments! I've updated the Github Actions to to use mamba, moved the commands to a single line as much as possible and enabled documentation caching.

Just a quick note to the reviewers: the documentation is currently broken while I'm working on the Github Actions to add some other functionality. Hopefully it'll be back by tonight.
