+++ .github/workflows/build_docs.yml
@@ -32,7 +32,7 @@ jobs:
       run: |
         micromamba install sphinx decorator h5py
         micromamba install -c conda-forge nbsphinx nbconvert
-        pip install sphinxcontrib-websupport sphinx_rtd_theme
+        pip install sphinxcontrib-websupport sphinx_rtd_theme sphinx-multiversion
 #    - uses: conda-incubator/setup-miniconda@v2
 #      with:
 #         activate-environment: xpsipy3
