### 2023-02-01 57dfb33
Adding a more sensible output directory format
