+++ docs/source/conf.py
@@ -69,15 +69,18 @@ html_sidebars = {
 
 
 #smv_rebuild_tags = False
-#smv_branch_whitelist = r'py3_multiversion_docs'
+smv_branch_whitelist = r"^main$"  # build main branch, and anything relating to documentation
 
 smv_remote_whitelist = r'^origin$'
 smv_prefer_remote_refs = True
 #smv_outputdir_format = '{ref.commit}' 
 
 ### OTHERS
-smv_released_pattern = r'^tags/v.*$'
+
+smv_released_pattern = r'.*tags.*'
 smv_outputdir_format = '{ref.name}'
+smv_remote_whitelist = r'^origin$'
+smv_latest_version = version
 #smv_tag_whitelist = r'^.*$'
 #smv_branch_whitelist = r'^.*$'
 #smv_branch_whitelist = r"^main$"  # build main branch, and anything relating to documentation
