### 2023-02-01 7be9faf
Added option to use either numpy allclose or the old alternative implementation printing the likelihood values. (#235)
