+++ docs/source/conf.py
@@ -210,20 +210,11 @@ texinfo_documents = [
 
 # -- Options for Multiversion Buid -----------------------------------------
 #smv_rebuild_tags = False
-smv_branch_whitelist = r"^main$"  # build main branch, and anything relating to documentation
+#smv_branch_whitelist = r"^main$"  # build main branch, and anything relating to documentation
 
 smv_remote_whitelist = r'^origin$'
 smv_prefer_remote_refs = True
-#smv_outputdir_format = '{ref.commit}' 
-
-### OTHERS
-
 smv_released_pattern = r'^tags/.*$'
 #smv_released_pattern = r'.*tags.*'
 smv_outputdir_format = '{ref.name}'
-smv_remote_whitelist = r'^origin$'
 smv_latest_version = release
-
-
-
-
