### 2023-02-01 26d8786
HPC installation instructions for Snellius+Helios updated for python3
