+++ docs/source/HPC_systems.rst
@@ -34,40 +34,29 @@ Load environment module and modify clean environment with Intel toolchain inform
 
 .. code-block:: bash
 
-    module load 2021
-    module load intel/2021a
+    module load 2022
+    module load intel/2022a
 
-Prepare python environment:
-
-.. code-block:: bash
-
-    module load Python/2.7.18-GCCcore-10.3.0-bare
-
-We will now install the various python packages we require. We use the module
-``/sw/arch/Centos8/EB_production/2021/modulefiles/lang/Python/2.7.18-GCCcore-10.3.0-bare.lua`` and its ``pip`` package manager to install packages
-locally in ``$HOME/.local/lib/python2.7/site-packages/`` if they are not
-installed globally or are outdated.
-
-Note that the python packages must be installed before pointing to intel compilers on Snellius to avoid child process reliability issues if using threads.
-`matplotlib` can be optionally installed for plotting purposes and is not necessary for modelling and inferencing.
+Prepare a conda environment for X-PSI:
 
 .. code-block:: bash
 
-    pip install --user Cython==0.29.24
-    pip install --user wrapt
-
-    pip install --user scipy
-    pip install --user numpy
-    pip install --user matplotlib
-    pip install --upgrade --user setuptools
-
-Set `PYTHONPATH`:
+    module load Anaconda3/2022.05
+    git clone https://github.com/xpsi-group/xpsi.git
+    cd xpsi
+    conda env create -f basic_environment.yml
+    conda init
+    
+For changes to take effect, close and re-open the current shell. After that load the modules again, and activate the environment:  
 
 .. code-block:: bash
 
-    export PYTHONPATH=$HOME/.local/lib/python2.7/site-packages/:$PYTHONPATH
-
-Point to Intel compilers:
+    module load 2022
+    module load intel/2022a
+    module load Anaconda3/2022.05
+    conda activate xpsi_py3
+    
+Next, we point to Intel compilers:
 
 .. code-block:: bash
 
@@ -87,33 +76,19 @@ To prepare MPI from ``$HOME``:
 
 .. code-block:: bash
 
-    wget https://bitbucket.org/mpi4py/mpi4py/downloads/mpi4py-3.0.0.tar.gz
-    tar -xvf mpi4py-3.0.0.tar.gz
-    cd mpi4py-3.0.0
-    python setup.py install --user
+    cd; wget https://bitbucket.org/mpi4py/mpi4py/downloads/mpi4py-3.1.4.tar.gz
+    tar zxvf mpi4py-3.1.4.tar.gz
+    cd mpi4py-3.1.4
+    python setup.py build   --mpicc=/sw/arch/RHEL8/EB_production/2022/software/impi/2021.6.0-intel-compilers-2022.1.0/mpi/2021.6.0/bin/mpicc
+    python setup.py install
 
 To test on the login node:
 
 .. code-block:: bash
 
-    mpiexec -n 8 python demo/helloworld.py
+    mpiexec -n 4 python demo/helloworld.py
 
-Do you see ranks 0 through 7 reporting for duty?
-
-.. note::
-
-    If faced with an error message ``Can't locate Switch.pm ...``, the `Switch` module can be loaded using the `CPAN` package manager as:
-
-    .. code-block:: bash
-
-        cpan
-
-    Follow the default instructions for the interactive cpan questions.
-
-    .. code-block:: bash
-
-        cpan[1]> install Switch
-        cpan[2]> exit
+Do you see ranks 0 through 3 reporting for duty?
 
 .. note::
 
@@ -141,7 +116,7 @@ Use the last command to check for the presence of shared objects.
 
 .. note::
 
[…]
