### 2023-02-01 82bb8e2
streamlined installation page
