+++ docs/source/install.rst
@@ -21,11 +21,11 @@ necessary pre-requisites on your local self-administered system.
 
 .. _dev_env:
 
-Python environment
-------------------
+Prerequisite Python packages
+----------------------------
 
 X-PSI was originally developed in Python 2.7 and was ported to Python 3 as of 
-X-PSI 2. We recommend creating a conda virtual environment with anaconda 3 as
+X-PSI 2. We recommend creating a conda virtual environment with anaconda3 as
 per instructions below so as to not disrupt your Python ecosystem.
 
 .. _basic_env:
@@ -33,57 +33,53 @@ per instructions below so as to not disrupt your Python ecosystem.
 Conda Environment
 ^^^^^^^^^^^^^^^^^
 
-In the source directory we provide a dependency file that installs
-the Python packages required for full functionality of X-PSI:
+In the source directory we provide a dependency file ``environment.yml`` that
+installs the Python packages required for full functionality of X-PSI. Its
+contents are:
 
-.. code-block:: 
-    name: xpsi2
+.. code-block:: bash
+
+    name: xpsi
     channels:
-      - defaults
-      - conda-forge
+        - defaults
+        - conda-forge
     dependencies:
-      - numpy
-      - cython
-      - matplotlib
-      - scipy
-      - wrapt
-      - pymultinest  # nested sampling
-      - mpi4py  # nested sampling
-      - getdist  # posterior KDE corner plotting
-      - h5py  # storage of X-ray signals computed from posterior samples
-      - nestcheck  # posterior error analysis, plotting, run combination, etc.
-      - fgivenx  # conditional posterior plotting; also required by nestcheck
+        - numpy
+        - cython
+        - matplotlib
+        - scipy
+        - wrapt
+        - pymultinest  # nested sampling
+        - mpi4py  # nested sampling
+        - getdist  # posterior KDE corner plotting
+        - h5py  # storage of X-ray signals computed from posterior samples
+        - nestcheck  # posterior error analysis, plotting, run combination, etc
+        - fgivenx  # conditional posterior plotting; also required by nestcheck
 
-The core packages required for likelihood functionality are:
+The core packages required for likelihood functionality are
+`numpy <https://docs.scipy.org/doc/numpy/index.html>`_,
+`cython <http://cython.readthedocs.io/en/latest>`_,
+`matplotlib <https://matplotlib.org/stable/index.html>`_,
+`scipy <https://docs.scipy.org/doc//scipy/index.html>`_, and
+`wrapt <https://wrapt.readthedocs.io/en/latest/>`_. Then, optional packages 
+required for nested sampling are
+`pymultinest <https://johannesbuchner.github.io/PyMultiNest/>`_ and
+`mpi4py <http://cython.readthedocs.io/en/latest>`_. Note that pyMultiNest 
+requires MultiNest, which may be installed later (refer to :ref:`multinest` in
+this guide). Finally, optional packages required for post-processing are
+`getdist <https://getdist.readthedocs.io/en/latest/>`_,
+`h5py <https://docs.h5py.org/en/stable/index.html>`_,
+`nestcheck <https://nestcheck.readthedocs.io/en/latest/>`_, and
+`fgivenx <https://fgivenx.readthedocs.io/en/latest/>`_.
 
-* `NumPy <https://docs.scipy.org/doc/numpy/index.html>`_
-* `Cython <http://cython.readthedocs.io/en/latest>`_
-* `Matplotlib <https://matplotlib.org/stable/index.html>`_
-* `Scipy <https://docs.scipy.org/doc//scipy/index.html>`_
-* `Wrapt <https://wrapt.readthedocs.io/en/latest/>`_
+In addition, some optional miscellaneous packages are:
 
-Then, optional packages required for nested sampling are: 
[…]
