+++ xpsi/utilities/ProjectionTool.py
@@ -17,6 +17,14 @@ from matplotlib import cm
 
 import xpsi
 
+"""
+The main idea of this tool it to create a 2D image showing the neutron star surface with its hot spots.
+The emisphere facing the observer has solid lines describing the hot spot; the emisphere facing opposite to the observer has dimmed colored lines compare dto what shown in the legend.
+The centers of a hot spots on the emisphere facing the observer are tagged with crosses; the ones correspondent to the hot spots on the opposite emisphere compare to the observer are tagged with cyrcles.
+Each hot spot is first created having one of the pole as center and then rotated to the right location.
+Further rotation are the necessary to account for the point of view.
+"""
+
 
 rc = {"font.family" : "serif",
       "mathtext.fontset" : "stix"}
@@ -35,12 +43,12 @@ rcParams['font.size'] = 14.0
 
 
 class IpyExit(SystemExit):
-    """Exit Exception for IPython.
-
+    """
+    Exit Exception for IPython.
     Exception temporarily redirects stderr to buffer.
     """
     def __init__(self):
-        # print("exiting")  # optionally print some message to stdout, too
+        print("exiting")  # optionally print some message to stdout, too
         # ... or do other stuff before exit
         sys.stderr = StringIO()
 
@@ -51,7 +59,17 @@ class IpyExit(SystemExit):
 
 
 def veneer(x, y, axes, lw=1.0, length=8):
-    """ Make the plots a little more aesthetically pleasing. """
+    """
+    GOAL: Make the plots a little more aesthetically pleasing. Author Thomas E. Riley
+    ##### INPUTS #####
+    x: sets a x-axis tick every x[0] (x values) and a number of the x-axis tick values every x[1] (x values)
+    y: sets a y-axis tick every y[0] (y values) and a number of the y-axis tick values every y[1] (y values)
+    axes: axes of plot
+    lw: tick line width in points
+    length: tick length in points
+    ##### OUTPUTS #####
+    -
+    """
     if x is not None:
         if x[1] is not None:
             axes.xaxis.set_major_locator(MultipleLocator(x[1]))
@@ -98,7 +116,14 @@ REF = dict(zip(ABBl,REFl))
 
 def transform(thetaR,phiR,V,phi0=0.0):
         """
-        
+        GOAL: rotate vector V to the right coordinates
+        ##### INPUTS #####
+        thetaR: desired colatitude of the vector (could be hot spot center) [rad]
+        phitR: desired phase of the vector (could be hot spot center) [cycles]
+        V: starting vector (could be defining the hot spot cyrcle)
+        phi0: phase of the point of view [cycle]
+        ##### OUTPUTS #####
+        Vout: rotated vector describing the hot spot
         """
         phi0 = phi0*np.pi*2.
 
@@ -235,9 +260,10 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
         """
         GOAL: checking compatibility between specified model and parameters
         ##### INPUTS #####
+        -
         ##### OUTPUTS #####
         hot_name: name of primary hot spot (according to our naming convention)
-        hot_name_s: name of the primary hot spot (according to our naming convention)
+        hot_name_s: name of the secondary hot spot (according to our naming convention)
         """
         hot_name = ''
         hot_name_s = ''
@@ -289,7 +315,7 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
[…]
