### 2023-02-01 a110425
Six changes for Py2 to Py3
