+++ examples/examples_modeling_tutorial/modules/CustomPrior_Beaming.py
@@ -1,4 +1,4 @@
-from __future__ import print_function, division
+
 
 import numpy as np
 import math
@@ -215,7 +215,7 @@ class CustomPrior(xpsi.Prior):
         p = list(p) # copy
 
         # used ordered names and values
-        ref = dict(zip(self.parameters.names, p))
+        ref = dict(list(zip(self.parameters.names, p)))
 
         # compactness ratio M/R_eq
         p += [gravradius(ref['mass']) / ref['radius']]
