+++ environment.yml
@@ -1,4 +1,4 @@
-name: xpsitest
+name: xpsi
 channels:
         - defaults
         - conda-forge
