+++ .github/workflows/build_docs.yml
@@ -66,7 +66,7 @@ jobs:
       shell: bash -l {0}
       run: |
         make versions
-        touch build/html/.nojekyll
+        #touch build/html/.nojekyll
         cp docs/docsite-index-redirect.html build/html/index.html
         ln -srf build/html/$(git tag -l --sort=creatordate |tail -1) build/html/latest  # auto-link the latest tag
 
