+++ .github/workflows/build_docs.yml
@@ -67,7 +67,7 @@ jobs:
       run: |
         make versions
         #touch build/html/.nojekyll
-        cp docs/docsite-index-redirect.html build/html/index.html
+        cp source/index-redirect.html build/html/index.html
         ln -srf build/html/$(git tag -l --sort=creatordate |tail -1) build/html/latest  # auto-link the latest tag
 
       working-directory: docs/
