### 2023-02-01 7151a4a
Added information on color code for hot spots
