+++ .github/workflows/build_docs.yml
@@ -51,7 +51,7 @@ jobs:
       run: |
         make versions
         #touch build/html/.nojekyll
-        mkdir build/html
+        #mkdir build/html
         cp source/index-redirect.html build/html/index.html
         ln -srf build/html/$(git tag -l --sort=creatordate |tail -1) build/html/latest  # auto-link the latest tag
 
