+++ docs/Makefile
@@ -9,6 +9,7 @@ SPHINXPROJ    = xpsi
 SOURCEDIR     = source
 BUILDDIR      = build
 SPHINXMULTIOPTS = -a
+MULTIBUILDDIR = build/html
 
 # Put it first so that "make" without argument is like "make help".
 help:
@@ -23,12 +24,5 @@ html:
 	@$(SPHINXBUILD) -M $@ "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
 
 versions:
-	echo @$(SPHINXMULTI) "$(SOURCEDIR)" "$(BUILDDIR)" "$(SPHINXOPTS)" $(0)
-	@$(SPHINXMULTI) "$(SOURCEDIR)" "$(BUILDDIR)" "$(SPHINXMULTIOPTS)" $(0)
+	@$(SPHINXMULTI) "$(SOURCEDIR)" "$(MULTIBUILDDIR)" "$(SPHINXMULTIOPTS)" $(0)
 
-# Catch-all target: route all unknown targets to Sphinx using the new
-# "make mode" option.  $(O) is meant as a shortcut for $(SPHINXOPTS).
-#%: Makefile
-#	@$(SPHINXBUILD) "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
-#%: Makefile
-#	@$(SPHINXBUILD) -M $@ "$(SOURCEDIR)" "$(BUILDDIR)" $(SPHINXOPTS) $(O)
