+++ xpsi/utilities/ProjectionTool.py
@@ -19,11 +19,11 @@ import xpsi
 
 """
 The main idea of this tool it to create a 2D image showing the neutron star surface with its hot spots.
-The emisphere facing the observer has solid lines describing the hot spot; the emisphere facing opposite to the observer has dimmed colored lines compare dto what shown in the legend.
-The centers of a hot spots on the emisphere facing the observer are tagged with crosses; the ones correspondent to the hot spots on the opposite emisphere compare to the observer are tagged with cyrcles.
-Each hot spot is first created having one of the pole as center and then rotated to the right location.
-Further rotation are the necessary to account for the point of view.
-Colors: omitting regions are marked in black; emitting regions are plotted with RdBu color map: from draker blue to dim blue, dim red and dark red going from hotter to colder emitting hot spots.
+The hemisphere facing the observer has solid lines describing the hot spot; the hemisphere opposite to the observer has dimmed colored lines compare to what is shown in the legend.
+The centers of a hot spot on the hemisphere facing the observer are tagged with crosses; the ones corresponding to the hot spots on the opposite hemisphere are tagged with cycles.
+Each hot spot is created with one of the poles as the center and then rotated to the right location.
+Further rotation is necessary to account for the point of view.
+Colors: omitting regions are marked in black; emitting regions are plotted with RdBu color map: dark blue -> light blue -> dim red -> dark red, moving from hotter to colder emitting hot spots.
 """
 
 
@@ -63,8 +63,8 @@ def veneer(x, y, axes, lw=1.0, length=8):
     """
     GOAL: Make the plots a little more aesthetically pleasing. Author Thomas E. Riley
     ##### INPUTS #####
-    x: sets a x-axis tick every x[0] (x values) and a number of the x-axis tick values every x[1] (x values)
-    y: sets a y-axis tick every y[0] (y values) and a number of the y-axis tick values every y[1] (y values)
+    x: sets a x-axis tick every x[0] value, and shows the number corresponding to the x-axis tick values every x[1] value
+    y: sets a y-axis tick every y[0] value, and shows the number corresponding to the y-axis tick values every y[1] value
     axes: axes of plot
     lw: tick line width in points
     length: tick length in points
@@ -121,7 +121,7 @@ def transform(thetaR,phiR,V,phi0=0.0):
         ##### INPUTS #####
         thetaR: desired colatitude of the vector (could be hot spot center) [rad]
         phitR: desired phase of the vector (could be hot spot center) [cycles]
-        V: starting vector (could be defining the hot spot cyrcle)
+        V: starting vector (could be defining the hot spot cycle)
         phi0: phase of the point of view [cycle]
         ##### OUTPUTS #####
         Vout: rotated vector describing the hot spot
@@ -712,9 +712,9 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
         symb: symbol plotted (unless signFlag is True and hot spot is visible: in that case marker is set as 'x')
         col: color of marker to plot the point of interest
         FlagLeg: True if we want to add the legend relative to the marker of interest to the plot
-        L1: label for legend if point of interest on the emisphere phasing observer
-        L2: label for legend if point of interest on the emisphere opposite to the observer
-        cL: color of marker for hot spot center in legend (only used of for this type of point different legend are used depending if the emisphere is facing or opposite to the observer)
[…]
