+++ .github/workflows/build_docs.yml
@@ -52,7 +52,7 @@ jobs:
         make versions
         touch build/html/.nojekyll
         cp source/index-redirect.html build/html/index.html
-        #ln -srf build/html/$(git tag -l --sort=creatordate |tail -1) build/html/latest  # auto-link the latest tag
+        ln -srf build/html/$(git tag -l --sort=creatordate |tail -1) build/html/latest  # auto-link the latest tag
 
       working-directory: docs/
     - name: Deploy to GitHub Pages
@@ -60,7 +60,7 @@ jobs:
       uses: crazy-max/ghaction-github-pages@v3
       with:
         target_branch: gh-pages
-        build_dir: docs/build/
+        build_dir: docs/build/html
         jekyll: false
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
