+++ .github/workflows/build_docs.yml
@@ -52,7 +52,7 @@ jobs:
         make versions
         touch build/html/.nojekyll
         cp source/index-redirect.html build/html/index.html
-        ln -srf build/html/$(git tag -l --sort=creatordate |tail -1) build/html/latest  # auto-link the latest tag
+        #ln -srf build/html/$(git tag -l --sort=creatordate |tail -1) build/html/latest  # auto-link the latest tag
 
       working-directory: docs/
     - name: Deploy to GitHub Pages
