### 2023-02-02 30e74b2
Adding options to execute CI tests manually
