+++ .github/workflows/ci_tests.yml
@@ -8,6 +8,7 @@ on:
     tags:
       - "*"
   pull_request:
+  workflow_dispatch:
 
 jobs:
   tests:
