### 2023-02-02 cb12e9d
trying to build both Py2 and Py3 docs at the same time
