+++ docs/source/HPC_systems.rst
@@ -3,13 +3,18 @@
 HPC systems
 ================
 
-The information provided in this page is for users who intend to work on High-Performance Computing (HPC) systems. These installation instructions are system-specific. X-PSI has already been used on different systems, for some of which, we provide the instructions below. This information may also be translated to other systems by users looking for guidance.
+The information provided in this page is for users who intend to work on 
+High-Performance Computing (HPC) systems. These installation instructions are 
+system-specific. X-PSI has already been used on different systems, for some of
+which we provide the instructions below. This information may also be
+translated to other systems by users looking for guidance.
 
 
 Snellius (SURF)
 -------------------
 
-`Snellius <https://servicedesk.surf.nl/wiki/display/WIKI/Snellius>`_ is the Dutch National Supercomputer.
+`Snellius <https://servicedesk.surf.nl/wiki/display/WIKI/Snellius>`_ is the 
+Dutch National Supercomputer.
 
 Installation
 ^^^^^^^^^^^^
@@ -30,7 +35,8 @@ To be additionally safe, run:
 
     module purge
 
-Load environment module and modify clean environment with Intel toolchain information:
+Load environment module and modify clean environment with Intel toolchain
+information:
 
 .. code-block:: bash
 
@@ -47,7 +53,8 @@ Prepare a conda environment for X-PSI:
     conda env create -f basic_environment.yml
     conda init
     
-For changes to take effect, close and re-open the current shell. After that load the modules again, and activate the environment:  
+For changes to take effect, close and re-open the current shell. After that
+load the modules again, and activate the environment:  
 
 .. code-block:: bash
 
@@ -64,7 +71,8 @@ Next, we point to Intel compilers:
     export CC=icc
     export CXX=icpc
 
-Below we explicitly specify flag arguments to select intel instruction sets that are compatible with the AMD processors present in Snellius.
+Below we explicitly specify flag arguments to select intel instruction sets
+that are compatible with the AMD processors present in Snellius.
 
 Load ``cmake`` module:
 
@@ -92,7 +100,8 @@ Do you see ranks 0 through 3 reporting for duty?
 
 .. note::
 
-    If MPI raises a warning about missing hydra process manager, run the following code-block:
+    If MPI raises a warning about missing hydra process manager, run the
+    following code-block:
 
     .. code-block:: bash
 
@@ -116,7 +125,10 @@ Use the last command to check for the presence of shared objects.
 
 .. note::
 
-    In case the Intel compilers on Snellius run into issues with Intel Math Kernal Library (MKL) due to static linkage, you can solve the problem by setting the appropriate paths to the environment variable for the pre-load libs:
+    In case the Intel compilers on Snellius run into issues with Intel Math
+    Kernel Library (MKL) due to static linkage, you can solve the problem by
+    setting the appropriate paths to the environment variable for the pre-load
+    libs:
 
     .. code-block:: bash
 
@@ -124,7 +136,8 @@ Use the last command to check for the presence of shared objects.
 
     Further details on MKL issues can be found in this `thread <https://community.intel.com/t5/Intel-oneAPI-Math-Kernel-Library/mkl-fails-to-load/m-p/1155538>`_
 
-We also need to the set the environment variable for library path to point at MultiNest:
+We also need to the set the environment variable for library path to point at
+MultiNest:
 
 .. code-block:: bash
 
@@ -153,8 +166,9 @@ Do you obtain parameter values and evidences?
     packages ``emcee`` and ``schwimmbad``. We assume the user can infer how to
     do this using the information above and on the :ref:`install` page.
 
[…]
