+++ examples/examples_fast/Modules/main.py
@@ -8,10 +8,10 @@ from xpsi.global_imports import gravradius
 
 import time
 
-from .CustomInstrument import CustomInstrument
-from .CustomSignal import CustomSignal
-from .CustomPhotosphere import CustomPhotosphere
-from .CustomPrior import CustomPrior
+from CustomInstrument import CustomInstrument
+from CustomSignal import CustomSignal
+from CustomPhotosphere import CustomPhotosphere
+from CustomPrior import CustomPrior
 
 
 # Data
