+++ xpsi/PostProcessing/_signalplotter.py
@@ -10,7 +10,7 @@ try:
     from tqdm.auto import trange as _range # detect if Jupyter notebook
 except ImportError:
     def _range(n, *args, **kwargs): # ignore extra (kw)args
-        return list(range(n))
+        return range(n)
 
 from ._global_imports import *
 
@@ -268,7 +268,7 @@ class SignalPlotter(PostProcessor):
 
                 # now restore the signals objects that were cached
                 cached = next(cache)
-                for key, value in list(cached.items()):
+                for key, value in cached.items():
                     try:
                         delattr(signal, key)
                     except AttributeError:
