### 2023-02-02 12fad3c
Revised citation and application files
