+++ docs/source/citation.rst
@@ -3,7 +3,6 @@
 Citation
 --------
 
-
 If X-PSI proves to be a useful tool for your work, please cite the project
 as a software acknowledgement, e.g.:
 
@@ -11,121 +10,40 @@ as a software acknowledgement, e.g.:
 
     X-PSI (\url{https://github.com/xpsi-group/xpsi.git}, \citet{xpsi})
 
-    @MISC{xpsi,
-        author = {{Riley}, Thomas Edward},
-        title = "{X-PSI: X-ray Pulse Simulation and Inference}",
-        keywords = {Software},
-        year = 2021,
-        month = feb,
-        eid = {ascl:2102.005},
-        pages = {ascl:2102.005},
-        archivePrefix = {ascl},
-        eprint = {2102.005},
-        adsurl = {https://ui.adsabs.harvard.edu/abs/2021ascl.soft02005R},
-        adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-    }
-
-A JOSS paper has been submitted to accompany the v1.0 release, and will eventually
-be the default paper to cite.  In the mean time some of the technical notes on
-X-PSI can be found in Chapter 3 of Thomas Riley's PhD thesis, which may also
-be cited.  
+and cite the X-PSI Journal of Open Source Software (JOSS) paper:
 
 .. code-block:: latex
 
-   @phdthesis{riley19,
-       author = {{Riley}, Thomas E.},
-       title = "{Neutron star parameter estimation from a NICER perspective}",
-       school = {University of Amsterdam},
-       year = 2019,
-       address = {https://hdl.handle.net/11245.1/aa86fcf3-2437-4bc2-810e-cf9f30a98f7a},
-       month = 12
-   }
-
-If you wish to point to the first published applications of the X-PSI package,
-then please cite :ref:`R19` and :ref:`R21`:
-
-.. code-block:: latex
-
-    @article{riley19,
-        author = {{Riley}, T.~E. and {Watts}, A.~L.
-                                 and {Bogdanov}, S.
-                                 and {Ray}, P.~S.
-                                 and {Ludlam}, R.~M.
-                                 and {Guillot}, S.
-                                 and {Arzoumanian}, Z.
-                                 and {Baker}, C.~L.
-                                 and {Bilous}, A.~V.
-                                 and {Chakrabarty}, D.
-                                 and {Gendreau}, K.~C.
-                                 and {Harding}, A.~K.
-                                 and {Ho}, W.~C.~G.
-                                 and {Lattimer}, J.~M.
-                                 and {Morsink}, S.~M.
-                                 and {Strohmayer}, T.~E.},
-        title = "{A NICER View of PSR J0030+0451: Millisecond Pulsar Parameter Estimation}",
-        doi = {10.3847/2041-8213/ab481c},
-        journal = {\apjl},
-        month = dec,
-        year = 2019,
-        volume = 887,
-        pages = {L21}
-    }
-
-    @article{riley21,
-	author = {{Riley}, Thomas E. 	and {Watts}, Anna L. 
-					and {Ray}, Paul S. 
-					and {Bogdanov}, Slavko 
-					and {Guillot}, Sebastien 
-					and {Morsink}, Sharon M. 
-					and {Bilous}, Anna V. 
-					and {Arzoumanian}, Zaven 
-					and {Choudhury}, Devarshi 
-					and {Deneva}, Julia S. 
-					and {Gendreau}, Keith C. 
-					and {Harding}, Alice K. 
-					and {Ho}, Wynn C.~G. 
-					and {Lattimer}, James M. 
-					and {Loewenstein}, Michael 
-					and {Ludlam}, Renee M. 
-					and {Markwardt}, Craig B. 
-					and {Okajima}, Takashi 
-					and {Prescod-Weinstein}, Chanda 
[…]
