+++ docs/source/applications.rst
@@ -3,81 +3,68 @@
 Applications
 ------------
 
-X-PSI was applied in the following studies, most recent listed first. 
-These may be useful for rough performance benchmarking and planning 
+X-PSI has been applied in the following studies. In addition to giving an
+idea of the scientific applications these may also
+be use useful for rough performance benchmarking and planning 
 of resource consumption.  For the most recent studies see the linked
 papers for full details of settings and resource consumption.
 
-If you have used X-PSI for a project and have time to summarise it here, please
+If you have used X-PSI for a project would like to link it here, please
 contact the X-PSI team and/or submit a pull-request on GitHub.
 
-.. _S22:
 
-`Salmi et al. 2022 (ApJ, 941 150)`__
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+Early X-PSI development
+***********************
 
-.. _ADS22: https://ui.adsabs.harvard.edu/abs/2022ApJ...941..150S/abstract
+X-PSI was developed by Thomas Riley as part of his PhD thesis work at the University of Amsterdam. 
+Chapter 3 of his `PhD thesis (Riley 2019) <https://hdl.handle.net/11245.1/aa86fcf3-2437-4bc2-810e-cf9f30a98f7a>`_ 
+provides a technical description of ``v0.1`` of X-PSI and contains
+results of some parameter recovery tests using synthetic data.  
 
-__ ADS22_
 
-`Salmi et al. (2022; hereafter S22)`__ used ``v0.7.10`` of X-PSI to model
-*NICER* observations of the rotation-powered millisecond X-ray pulsar PSR J0740+6620 using *NICER*
-background estimates.
+NICER papers
+************
 
-__ ADS22_
+X-PSI has been used in several Pulse Profile Modeling analysis papers by the *NICER* team. These are typically published with a Zenodo repository containing data, model files, X-PSI scripts, samples and post-processing notebooks to enable full reproduction and independent analysis. 
 
-See also the associated `Zenodo repository`__.
+**Salmi et al. 2022**
+
+`Salmi et al. 2022 (ApJ, 941, 150) <https://ui.adsabs.harvard.edu/abs/2022ApJ...941..150S/abstract>`_ used  ``v0.7.10`` of X-PSI to model *NICER* observations of the rotation-powered millisecond X-ray pulsar PSR J0740+6620 using *NICER* background estimates.  See also the associated `Zenodo repository`__.
 
 .. _Zenodo22: https://doi.org/10.5281/zenodo.6827536
 __ Zenodo22_
 
 
-.. _R21:
+**Riley et al. 2021**
 
-`Riley et al. 2021 (ApJL, 918, L27)`__
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. _ADS21: https://ui.adsabs.harvard.edu/abs/2021ApJ...918L..27R/abstract
-
-__ ADS21_
-
-`Riley et al. (2021; hereafter R21)`__ used ``v0.7.6`` of X-PSI to model
-*NICER* observations of the rotation-powered millisecond X-ray pulsar PSR J0740+6620.
-
-__ ADS21_
-
-See also the associated `Zenodo repository`__.
+`Riley et al. 2021 (ApJL, 918, L27) <https://ui.adsabs.harvard.edu/abs/2021ApJ...918L..27R/abstract>`_ used ``v0.7.6`` of X-PSI to model *NICER* observations of the rotation-powered millisecond X-ray pulsar PSR J0740+6620. See also the associated `Zenodo repository`__.
 
 .. _Zenodo21: https://doi.org/10.5281/zenodo.4697624
 __ Zenodo21_
 
[…]
