### 2023-02-02 d9ff6fd
implemented pr requested changes in install.rst
