### 2023-02-03 8266b11
Six changes for Py2 to Py3 (#244)
* Six changes for Py2 to Py3

* Corrections to six changes
