+++ examples/examples_fast/Modules/CustomInstrument.py
@@ -1,4 +1,4 @@
-from __future__ import print_function, division
+
 
 import numpy as np
 import math
