### 2023-02-03 711862e
Update install.rst
- some clarifications
- simplified instructions for Windows
