### 2023-02-03 a7ef914
A few more general py2 doc updates
