+++ docs/source/todo.rst
@@ -4,5 +4,5 @@ Future
 ------
 
 All good things come to an end and sadly there is no future for the Python 2 variant
-of X-PSI.  Please return to the main X-PSI page and switch to the Python 3 (X-PSI v>2.0)
-variant and see what the future holds! 
+of X-PSI.  Please return to the `main X-PSI page <https://xpsi-group.github.io/xpsi/>`_
+ and switch to the Python 3 (X-PSI v>2.0) variant to see what the future holds! 
