### 2023-02-03 9dcd71f
added basic functionality tests for hotregion and hotregions
