### 2023-02-03 697ddbe
Module generator tutorial added.
