+++ docs/source/Module_generator_tutorial.rst
@@ -1,0 +1,453 @@
+.. _module_generator_tutorial:
+
+Module generator tutorial
+=========================
+
+To develop models in X-PSI, users are recommended to to create custom derived classes for the different X-PSI model
+components (e.g. ``Prior``, ``Photosphere``, etc.) that require further modification. Examples of such `Custom` classes
+can be found in :ref:`./Modeling.ipynb`. These classes are preferably to be written as distinct modules within a project
+directory rather than in a notebook as is done in :ref:`./Modeling.ipynb`. The modules can then be imported by a `Main`
+script, which can be executed using an MPI command in a shell to exploit parallelism for resource-intensive likelihood
+evaluations.
+
+The structure of these `Main` and `Custom` modules are largely the same for most user-defined models. It's usually the
+scope of these models (e.g. hotspot models, required data and instrument files, prior definitions, etc.) that are
+subject to change. Therefore, to simplify the process and avoid the user from having to create these modules everytime
+for a new model, we have created the program ``module_generator`` which takes in arguments to define the scope of the
+model. Below, we provide instructions on how to use it.
+
+Create a symlink of the module generator into your project directory. You can run the help command as shown below to
+obtain details regarding its usage.
+
+.. code-block:: bash
+
+    ln -s <path/to/xpsi>/xpsi/module_generator.py ./
+    python module_generator.py -h
+
+The output of the above code block is:
+
+.. code-block:: bash
+
+    /=============================================\
+    | X-PSI: X-ray Pulse Simulation and Inference |
+    |---------------------------------------------|
+    |                Version: 2.0.0               |
+    |---------------------------------------------|
+    |      https://xpsi-group.github.io/xpsi      |
+    \=============================================/
+
+    Imported GetDist version: 1.4
+    Imported nestcheck version: 0.2.1
+    Parsing configuration file...
+    usage: module_generator.py [-h] [--generate-config-file]
+                               [--telescope TELESCOPE] [--instrument INSTRUMENT]
+                               [--source SOURCE] --frequency FREQUENCY
+                               [--model MODEL] --hot-region-model
+                               {ST,CST,EST,PST,CDT,EDT,PDT}
+                               [--antipodal-reflection-symmetry]
+                               [--break-hot-region-exchange-degeneracy-with BREAK_HOT_REGION_EXCHANGE_DEGENERACY_WITH]
+                               [--is-antiphased IS_ANTIPHASED] [--prefix PREFIX]
+                               [--hot-atmosphere-model HOT_ATMOSPHERE_MODEL]
+                               [--hot-atmosphere-load]
+                               [--elsewhere-atmosphere-model ELSEWHERE_ATMOSPHERE_MODEL]
+                               [--elsewhere-atmosphere-load]
+                               [--attenuation-model ATTENUATION_MODEL]
+                               [--background-model] [--background-shared-instance]
+                               [--background-shared-class]
+                               [--background-parameters [BACKGROUND_PARAMETERS ...]]
+                               [--print-MPI-rank] [--config-path CONFIG_PATH]
+                               [--module-directory-path MODULE_DIRECTORY_PATH]
+                               [--main-module MAIN_MODULE]
+                               [--custom-signal-module CUSTOM_SIGNAL_MODULE]
+                               [--custom-instrument-module CUSTOM_INSTRUMENT_MODULE]
+                               [--custom-photosphere-module CUSTOM_PHOTOSPHERE_MODULE]
+                               [--custom-interstellar-module CUSTOM_INTERSTELLAR_MODULE]
+                               [--custom-prior-module CUSTOM_PRIOR_MODULE]
+                               [--custom-background-module CUSTOM_BACKGROUND_MODULE]
+
[…]
