+++ examples/examples_module_generator/generate.ini
@@ -1,0 +1,49 @@
+##----------------------------##
+## telescope instrument flags ##
+##----------------------------##
+--telescope=NICER
+--instrument=XTI
+
+
+##---------------------##
+## target source flags ##
+##---------------------##
+--source=PSR J0030+0415
+--frequency=205
+
+
+##-------------##
+## model flags ##
+##-------------##
+--model=CST+PDT
+--hot-region-model=CST
+--hot-region-model=PDT
+--is-antiphased=False
+--is-antiphased=False
+--prefix=p
+--prefix=s
+--hot-atmosphere-model=NSX-H
+--hot-atmosphere-load
+--attenuation-model=tbnew
+
+
+
+
+
+##---------------------##
+## miscellaneous flags ##
+##---------------------##
+--print-MPI-rank
+
+
+##-------------##
+## write flags ##
+##-------------##
+--config-path=./config.ini
+--module-directory-path=./_auto_modules
+--main-module=main
+--custom-signal-module=CustomSignal
+--custom-instrument-module=CustomInstrument
+--custom-photosphere-module=CustomPhotosphere
+--custom-interstellar-module=CustomInterstellar
+--custom-prior-module=CustomPrior
