+++ examples/examples_module_generator/generate.ini
@@ -8,7 +8,7 @@
 ##---------------------##
 ## target source flags ##
 ##---------------------##
---source=PSR J0030+0415
+--source=PSR J0030+0451
 --frequency=205
 
 
