### 2023-02-03 0d23844
Updated `index.rst` to reflect module generator
