+++ docs/source/index.rst
@@ -36,6 +36,7 @@
    Post-processing
    Emitting_patterns_2Dprojection
    Importance_sample
+   Module_generator_tutorial
    Example_script_and_modules
    Example_job
 
