### 2023-02-03 6070cfa
245 add documentation projection tool (#258)
* changed default lenguage; added ipynb for tutorials on ProjectionTool and Importance sampling; added tutorials in index.rst

* Starting with the ipynb on the projection tool

* Completed projection tool tutirial; some bug fixes necessary to the ProjectionTool

* Adding link to other tutorial and re run Projection Tool Tutorial without exits

* fixing some typos in the Projection tool tutorial

* Fixing further typo

* Updated `index.rst` to reflect module generator

---------

Co-authored-by: Serena Vinciguerra <s.vinciguerra@uva.nl>
Co-authored-b
[…]
