+++ xpsi/utilities/ProjectionTool.py
@@ -218,12 +218,11 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
         elif (asymFlag and ((P_name in params) or (C_name in params))):
             print ("WARNING! there are info for a secondary hot spot that will not being used")
             
-        
         if ('DT' in hot_name):
             ### check cede
             T_name = HStag+REF['CT']
             R_name = HStag+REF['CR']
-            if ('C' in hot_name) and not((T_name in params) or (R_name in params)):
+            if ('C' in hot_name) and (not(T_name in params) or not(R_name in params)):
                 print ("ERROR! Double temperature (DT) models require the definition of cede properties")
                 raise IpyExit
             C_name = HStag+REF['CT']
@@ -231,30 +230,46 @@ def plot_projection_general(dictVp, model, POV = "", ThetaDisplay = "",antiphase
             
             if ('C' in hot_name) and ((C_name in params) or (A_name in params)):
                 print ("WARNING! there are info for a complex geometry, but they are not being used")
-            if (('P' in hot_name) or ('E' in hot_name)) and not((C_name in params) or (C_name in params)):
+            if (('P' in hot_name) or ('E' in hot_name)) and (not(C_name in params) or not(C_name in params)):
                 print ("ERROR! Protruding and Eccentric models require the definition of cede colatitudes and phases")
                 raise IpyExit
-            if not(('C' in hot_name) ^ ('E' in hot_name) ^ ('P'in hot_name)):
+            if not(('C' in hot_name) ^ ('E' in hot_name) ^ ('P' in hot_name)):
                 print ("ERROR! Double temperature models require the setting if concentric (C, hot spot name = CDT), eccentric (E, hot spot name = EDT), protruding (P, hot spot name = PDT), current name is",hot_name)
                 raise IpyExit
-        
-        elif ('ST' in hot_name) and (('C' in hot_name) ^ ('E' in hot_name) ^ ('P 'in hot_name)):
+            if (('C' in hot_name) or ('E' in hot_name)) and dictVp[R_name]<dictVp[HStag+REF['SR']]:
+                if ('C' in hot_name):
+                    print ("WARNING: ceding component is completely masked by the superseding one")
+                else:
+                    print ("WARNING: hot spot decleared ECCENTRIC, but parameters will render a PROTRUDING configuration")
+            
+        elif ('ST' in hot_name) and (('C' in hot_name) ^ ('E' in hot_name) ^ ('P' in hot_name)):
             R_name = HStag+REF['OR']
             if ('C' in hot_name) and not((R_name in params)):
                 print ("ERROR! models including an omission component require the definition of omit properties")
                 raise IpyExit
             C_name = HStag+REF['OC']
             A_name = HStag+REF['OA']
+            if (('C' in hot_name) or ('E' in hot_name)) and dictVp[R_name]>dictVp[HStag+REF['SR']]:
+                if ('C' in hot_name):
+                    print ("WARNING: emitting component is completely masked by the omitting one")
+                else:
+                    print ("WARNING: hot spot decleared ECCENTRIC, but parameters will render a PROTRUDING configuration")
+                
             if (not ('P' in hot_name) and not('E' in hot_name)) and ((C_name in params) or (A_name in params)):
[…]
