### 2023-02-03 53687d2
Fixes #197; Intensity units corrected and unit conversion factors directly read from global variables.
