### 2023-02-03 41d36b1
Documentation example for surface radiation field updated.
