+++ xpsi/global_imports.py
@@ -73,7 +73,7 @@ _kpc = 3.08567758e19
 _keV = 1.60217662e-16
 _GM = 0.5 * 2.95325024e3 # Solar Schwarzschild gravitational radius in SI
 _G = 6.6740831e-11
-_h_keV = 4.135667662e-18 #Conversion from /Hz to /keV units
+_h_keV = 4.135667662e-18 #Conversion from Hz^{-1} and keV^{-1} units
 _h = 6.62607004e-34
 _dpr = 180.0 / _pi
 _km = 1.0e3
