### 2023-02-03 0559fca
Bug in synthetic data seed setting fixed when seed=0
