### 2023-02-09 a485358
Fixes #197; Intensity unit corrections and comments. (#265)
* Fixes #197; Intensity units corrected and unit conversion factors directly read from global variables.

* Documentation example for surface radiation field updated.

* Syntax for inverse units changed.
