### 2023-02-09 #4977 @adonath
Thanks a lot @drannawatts!

The manuscript reads nicely and contains all the required information on X-PSI. The frequent use of parentheses somewhat distracted the reading flow. In many cases I considered the information in parentheses more than optional. So I would suggest to include the information directly in the sentence instead. Here is a summary of mostly minor comments on the manuscript from my side:

- Stick to title case consistently? So change "package", "neutron" and "star" to upper case. In case you do, please also adapt the section title. Otherwise change "Pulse Simulation and Inference" to lower case for consistency.
- You nowhere mention, what the X-PSI acronym stands for
- Line 11: you introduce the equivalency of "rotationally-modified" and "pulsed", which you later repeat in line 16. Maybe change to "rotationally modulated" in line 11 and just use "pulsed" in line 16.
- Line 19: I think replacing the parentheses just by a comma improves the reading flow.
- Line 20: same here, remove parentheses and just use "...of the star, which determines the..."
- Line 26: could add an example "...future telescopes, such as ..."
- Line 36: again replace parentheses e.g. by using "..., including likelihood and prior functions, ..."
- Line 40: here I understand that  "(animated)" would be equivalent to "time- and energy-resolved". This might be true if time-resolution is always achieved by animation, which is probably not the case. Maybe just skip parentheses? So images are time- and energy-resolved **and** animated? Which I think is the case...
- Line 53: I would again remove the parentheses. Using parentheses makes the contained information appear unimportant, which I think is not the case.

Once the comments are clarified or addressed, you can go ahead and tag a new version of X-PSI. This version should be uploaded to a software archive, such as Zenodo. Please report the version tag and DOI here in the issue after.
