### 2023-02-10 edbe49f
Checking out branch python2 mentioned in main installation page.
