### 2023-02-10 2e54a64
Git checkout python2 added to the HPC instructions as well.
