### 2023-02-11 0cd43d3
Updates to JOSS paper on main branch
