+++ joss/xpsijoss.bib
@@ -145,22 +145,42 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@INPROCEEDINGS{Watts2019,
-       author = {{Watts}, Anna L.},
-        title = "{Constraining the neutron star equation of state using pulse profile modeling}",
-     keywords = {Astrophysics - High Energy Astrophysical Phenomena, Astrophysics - Solar and Stellar Astrophysics, General Relativity and Quantum Cosmology, Nuclear Theory},
-    booktitle = {Xiamen-CUSTIPEN Workshop on the Equation of State of Dense Neutron-Rich Matter in the Era of Gravitational Wave Astronomy},
+
+
+@ARTICLE{Watts2019,
+       author = {{Watts}, Anna L. and {Yu}, WenFei and {Poutanen}, Juri and {Zhang}, Shu and {Bhattacharyya}, Sudip and {Bogdanov}, Slavko and {Ji}, Long and {Patruno}, Alessandro and {Riley}, Thomas E. and {Bakala}, Pavel and {Baykal}, Altan and {Bernardini}, Federico and {Bombaci}, Ignazio and {Brown}, Edward and {Cavecchi}, Yuri and {Chakrabarty}, Deepto and {Chenevez}, J{\'e}r{\^o}me and {Degenaar}, Nathalie and {Del Santo}, Melania and {Di Salvo}, Tiziana and {Doroshenko}, Victor and {Falanga}, Maurizio and {Ferdman}, Robert D. and {Feroci}, Marco and {Gambino}, Angelo F. and {Ge}, MingYu and {Greif}, Svenja K. and {Guillot}, Sebastien and {Gungor}, Can and {Hartmann}, Dieter H. and {Hebeler}, Kai and {Heger}, Alexander and {Homan}, Jeroen and {Iaria}, Rosario and {Zand}, Jean in't. and {Kargaltsev}, Oleg and {Kurkela}, Aleksi and {Lai}, XiaoYu and {Li}, Ang and {Li}, XiangDong and {Li}, ZhaoSheng and {Linares}, Manuel and {Lu}, FangJun and {Mahmoodifar}, Simin and {M{\'e}ndez}, Mariano and {Coleman Miller}, M. and {Morsink}, Sharon and {N{\"a}ttil{\"a}}, Joonas and {Possenti}, Andrea and {Prescod-Weinstein}, Chanda and {Qu}, JinLu and {Riggio}, Alessandro and {Salmi}, Tuomo and {Sanna}, Andrea and {Santangelo}, Andrea and {Schatz}, Hendrik and {Schwenk}, Achim and {Song}, LiMing and {{\v{S}}r{\'a}mkov{\'a}}, Eva and {Stappers}, Benjamin and {Stiele}, Holger and {Strohmayer}, Tod and {Tews}, Ingo and {Tolos}, Laura and {T{\"o}r{\"o}k}, Gabriel and {Tsang}, David and {Urbanec}, Martin and {Vacchi}, Andrea and {Xu}, RenXin and {Xu}, YuPeng and {Zane}, Silvia and {Zhang}, GuoBao and {Zhang}, ShuangNan and {Zhang}, WenDa and {Zheng}, ShiJie and {Zhou}, Xia},
+        title = "{Dense matter with eXTP}",
+      journal = {Science China Physics, Mechanics, and Astronomy},
+     keywords = {neutron, X-rays, dense matter, equation of state, Astrophysics - High Energy Astrophysical Phenomena},
          year = 2019,
-       series = {American Institute of Physics Conference Series},
-       volume = {2127},
-        month = jul,
-          eid = {020008},
[…]
