### 2023-02-11 ea3fb42
Python2 installation update (#279)
* Checking out branch python2 mentioned in main installation page.

* Git checkout python2 added to the HPC instructions as well.

* PyMultiNest Python2 compatible commit mentioned.

* Corrected Helios installation instruction to include GSL installation for both methods.
