### 2023-02-11 #4977 @drannawatts
Comments are all addressed, thank you @adonath !

New version of X-PSI is tagged (v1.2.2) and uploaded to Zenodo at https://zenodo.org/record/7632629
The doi is  10.5281/zenodo.7632629
