### 2023-02-13 8697041
updating journal names and citation syntax
