### 2023-02-13 #4977 @dfm
@drannawatts — I've submitted a PR with some minor edits, but I have one other request before the final processing: Can you update the metadata for [the Zenodo archive](https://zenodo.org/record/7632629) so that the title and author list _exactly match the paper_? There should be an "edit" button at the top right of that page.
