### 2023-02-14 #4977 @dfm
@drannawatts et al. — Your paper is now accepted and published in JOSS! :zap::rocket::boom:
